Extract Model Fitted Values for intPGOcc Object

Method for extracting model fitted values and detection probability values from a fitted single-species integrated occupancy (intPGOcc) model.

# S3 method for intPGOcc
fitted(object, ...)

Arguments

object

object of class intPGOcc.

...

currently no additional arguments

Details

A method to the generic fitted function to extract fitted values and detection probability values for fitted model objects of class intPGOcc.

Value

A list comprised of

y.rep.samples

A list of three-dimensional numeric arrays of fitted values for each individual data source for use in Goodness of Fit assessments.

p.samples

A list of three-dimensional numeric arrays of detection probability values.