Function for Fitting Single Species Integrated Occupancy Models Using Polya-Gamma Latent Variables

Function for fitting single species integrated occupancy models using Polya-Gamma latent variables. Data integration is done using a joint likelihood framework, assuming distinct detection models for each data source that are each conditional on a single latent occurrence process.

intPGOcc(occ.formula, det.formula, data, inits, priors, n.samples, 
         n.omp.threads = 1, verbose = TRUE, n.report = 1000, 
         n.burn = round(.10 * n.samples), n.thin = 1, 
         k.fold, k.fold.threads = 1, k.fold.seed, 
         k.fold.data, ...
)

Arguments

occ.formula
a symbolic description of the model to be fit for the occurrence portion of the model using R's model syntax. Only right-hand side of formula is specified. See example below.

det.formula
a list of symbolic descriptions of the models to be fit for the detection portion of the model using R's model syntax for each data set. Each element in the list is a formula for the detection model of a given data set. Only right-hand side of formula is specified. See example below.

data
a list containing data necessary for model fitting. Valid tags are y, occ.covs, det.covs, and sites. y is a list of matrices or data frames for each data set used in the integrated model. Each element of the list has first dimension equal to the number of sites with that data source and second dimension equal to the maximum number of replicates at a given site. occ.covs is a matrix or data frame containing the variables used in the occupancy portion of the model, with the number of rows being the number of sites with at least one data source for each column (variable). det.covs is a list of variables included in the detection portion of the model for each data source. det.covs should have the same number of elements as y, where each element is itself a list. Each element of the list for a given data source is a different detection covariate, which can be site-level or observational-level. Site-level covariates are specified as a vector with length equal to the number of observed sites of that data source, while observation-level covariates are specified as a matrix or data frame with the number of rows equal to the number of observed sites of that data source and number of columns equal to the maximum number of replicates at a given site.

inits
a list with each tag corresponding to a parameter name. Valid tags are z, beta, and alpha. The value portion of tags z and beta is the parameter's initial value. The tag alpha is a list comprised of the initial values for the detection parameters for each data source. Each element of the list should be a vector of initial values for all detection parameters in the given data source or a single value for each data source to assign all parameters for a given data source the same initial value. See priors description for definition of each parameter name.

priors
a list with each tag corresponding to a parameter name. Valid tags are beta.normal and alpha.normal. Occurrence (beta) and detection (alpha) regression coefficients are assumed to follow a normal distribution. For beta hyperparameters of the normal distribution are passed as a list of length two with the first and second elements corresponding to the mean and variance of the normal distribution, which are each specified as vectors of length equal to the number of coefficients to be estimated or of length one if priors are the same for all coefficients. For the detection coefficients alpha, the mean and variance hyperparameters are themselves passed in as lists, with each element of the list corresponding to the specific hyperparameters for the detection parameters in a given data source. If not specified, prior means are set to 0 and prior variances set to 2.72.

n.samples
the number of posterior samples to collect.

n.omp.threads
a positive integer indicating the number of threads to use for SMP parallel processing. The package must be compiled for OpenMP support. For most Intel-based machines, we recommend setting n.omp.threads up to the number of hypterthreaded cores. Note, n.omp.threads > 1 might not work on some systems.

verbose
if TRUE, messages about data preparation, model specification, and progress of the sampler are printed to the screen. Otherwise, no messages are printed.

n.report
the interval to report MCMC progress.

n.burn
the number of samples out of the total n.samples to discard as burn-in. By default, the first 10% of samples is discarded.

n.thin
the thinning interval for collection of MCMC samples. The thinning occurs after the n.burn samples are discarded. Default value is set to 1.

k.fold
specifies the number of k folds for cross-validation. If not specified as an argument, then cross-validation is not performed and k.fold.threads and k.fold.seed are ignored. In k-fold cross-validation, the data specified in data is randomly partitioned into k equal sized subsamples. Of the k subsamples, k - 1 subsamples are used to fit the model and the remaining k samples are used for prediction. The cross-validation process is repeated k times (the folds). As a scoring rule, we use the model deviance as described in Hooten and Hobbs (2015). Cross-validation is performed after the full model is fit using all the data. Cross-validation results are reported in the k.fold.deviance object in the return list.

k.fold.threads
number of threads to use for cross-validation. If k.fold.threads > 1 parallel processing is accomplished using the foreach and doParallel packages. Ignored if k.fold is not specified.

k.fold.seed
seed used to split data set into k.fold parts for k-fold cross-validation. Ignored if k.fold is not specified.

k.fold.data
an integer specifying the specific data set to hold out values from. If not specified, data from all data set locations will be incorporated into the k-fold cross-validation.

...
currently no additional arguments

Note

Some of the underlying code used for generating random numbers from the Polya-Gamma distribution is taken from the pgdraw package written by Daniel F. Schmidt and Enes Makalic. Their code implements Algorithm 6 in PhD thesis of Jesse Bennett Windle (2013) https://repositories.lib.utexas.edu/handle/2152/21842.

References

Polson, N.G., J.G. Scott, and J. Windle. (2013) Bayesian Inference for Logistic Models Using Polya-Gamma Latent Variables. Journal of the American Statistical Association, 108:1339-1349.

Hooten, M. B., and Hobbs, N. T. (2015). A guide to Bayesian model selection for ecologists. Ecological monographs, 85(1), 3-28.

Finley, A. O., Datta, A., and Banerjee, S. (2020). spNNGP R package for nearest neighbor Gaussian process models. arXiv preprint arXiv:2001.09111.

Author

Jeffrey W. Doser doserjef@msu.edu,
Andrew O. Finley finleya@msu.edu

Value

An object of class intPGOcc that is a list comprised of:

beta.samples

a coda object of posterior samples for the occupancy regression coefficients.

alpha.samples

a coda object of posterior samples for the detection regression coefficients for all data sources.

z.samples

a coda object of posterior samples for the latent occupancy values

psi.samples

a coda object of posterior samples for the latent occupancy probability values

y.rep.samples

a list of three-dimensional arrays of fitted values for each data source for use in Goodness of Fit assessments.

run.time

execution time reported using proc.time().

k.fold.deviance

scoring rule (deviance) from k-fold cross-validation. A separate deviance value is returned for each data source. Only included if k.fold is specified in function call. Only a single value is returned if k.fold.data is specified.

The return object will include additional objects used for subsequent prediction and/or model fit evaluation.
Examples

set.seed(1008)

# Simulate Data -----------------------------------------------------------
J.x <- 15
J.y <- 15
J.all <- J.x * J.y
# Number of data sources.
n.data <- 4
# Sites for each data source. 
J.obs <- sample(ceiling(0.2 * J.all):ceiling(0.5 * J.all), n.data, replace = TRUE)
# Replicates for each data source.
n.rep <- list()
for (i in 1:n.data) {
  n.rep[[i]] <- sample(1:4, size = J.obs[i], replace = TRUE)
}
# Occupancy covariates
beta <- c(0.5, 1)
p.occ <- length(beta)
# Detection covariates
alpha <- list()
for (i in 1:n.data) {
  alpha[[i]] <- runif(2, -1, 1)
}
p.det.long <- sapply(alpha, length)
p.det <- sum(p.det.long)

# Simulate occupancy data. 
dat <- simIntOcc(n.data = n.data, J.x = J.x, J.y = J.y, J.obs = J.obs, 
     n.rep = n.rep, beta = beta, alpha = alpha, sp = FALSE)

y <- dat$y
X <- dat$X.obs
X.p <- dat$X.p
sites <- dat$sites

# Package all data into a list
occ.covs <- X[, 2, drop = FALSE]
colnames(occ.covs) <- c('occ.cov')
det.covs <- list()
# Add covariates one by one
det.covs[[1]] <- list(det.cov.1.1 = X.p[[1]][, , 2]) 
det.covs[[2]] <- list(det.cov.2.1 = X.p[[2]][, , 2]) 
det.covs[[3]] <- list(det.cov.3.1 = X.p[[3]][, , 2]) 
det.covs[[4]] <- list(det.cov.4.1 = X.p[[4]][, , 2]) 
data.list <- list(y = y, 
      occ.covs = occ.covs,
      det.covs = det.covs, 
      sites = sites)

J <- length(dat$z.obs)
# Initial values
inits.list <- list(alpha = list(0, 0, 0, 0), 
          beta = 0, 
          z = rep(1, J))
# Priors
prior.list <- list(beta.normal = list(mean = 0, var = 2.72), 
       alpha.normal = list(mean = list(0, 0, 0, 0), 
                     var = list(2.72, 2.72, 2.72, 2.72)))
n.samples <- 5000
out <- intPGOcc(occ.formula = ~ occ.cov, 
                det.formula = list(f.1 = ~ det.cov.1.1, 
                       f.2 = ~ det.cov.2.1, 
                       f.3 = ~ det.cov.3.1, 
                       f.4 = ~ det.cov.4.1), 
    data = data.list,
    inits = inits.list,
    n.samples = n.samples, 
    priors = prior.list, 
    n.omp.threads = 1, 
    verbose = TRUE, 
    n.report = 1000, 
    n.burn = 1000, 
    n.thin = 1)
#> ----------------------------------------
#> 	Preparing the data
#> ----------------------------------------
#> ----------------------------------------
#> 	Model description
#> ----------------------------------------
#> Integrated Occupancy Model with Polya-Gamma latent
#> variable fit with 167 sites.
#> 
#> Integrating 4 occupancy data sets.
#> 
#> Number of MCMC samples: 5000
#> Burn-in: 1000 
#> Thinning Rate: 1 
#> Total Posterior Samples: 4000 
#> 
#> 
#> Source compiled with OpenMP support and model fit using 1 thread(s).
#> 
#> Sampling ... 
#> Sampled: 1000 of 5000, 20.00%
#> -------------------------------------------------
#> Sampled: 2000 of 5000, 40.00%
#> -------------------------------------------------
#> Sampled: 3000 of 5000, 60.00%
#> -------------------------------------------------
#> Sampled: 4000 of 5000, 80.00%
#> -------------------------------------------------
#> Sampled: 5000 of 5000, 100.00%

summary(out)
#> 
#> Call:
#> intPGOcc(occ.formula = ~occ.cov, det.formula = list(f.1 = ~det.cov.1.1, 
#>     f.2 = ~det.cov.2.1, f.3 = ~det.cov.3.1, f.4 = ~det.cov.4.1), 
#>     data = data.list, inits = inits.list, priors = prior.list, 
#>     n.samples = n.samples, n.omp.threads = 1, verbose = TRUE, 
#>     n.report = 1000, n.burn = 1000, n.thin = 1)
#> 
#> Chain Information:
#> Total samples: 5000
#> Burn-in: 1000
#> Thin: 1
#> Total Posterior Samples: 4000
#> 
#> Occurrence: 
#>               2.5%    25%    50%    75%  97.5%
#> (Intercept) 0.0866 0.3268 0.4623 0.5922 0.8714
#> occ.cov     0.5840 0.8384 0.9848 1.1361 1.4573
#> 
#> Data source 1 Detection: 
#>                2.5%     25%     50%     75%   97.5%
#> (Intercept)  0.1048  0.4033  0.5816  0.7507  1.0894
#> det.cov.1.1 -1.2405 -0.8394 -0.6480 -0.4587 -0.1275
#> 
#> Data source 2 Detection: 
#>                2.5%     25%     50%     75%   97.5%
#> (Intercept)  0.0628  0.3839  0.5535  0.7288  1.0894
#> det.cov.2.1 -2.9857 -2.3789 -2.0899 -1.8254 -1.3587
#> 
#> Data source 3 Detection: 
#>                2.5%     25%     50%     75%   97.5%
#> (Intercept) -0.9371 -0.6366 -0.4781 -0.3230 -0.0135
#> det.cov.3.1 -0.0133  0.3086  0.4818  0.6566  1.0029
#> 
#> Data source 4 Detection: 
#>                2.5%    25%    50%   75%  97.5%
#> (Intercept) -0.0375 0.2617 0.4276 0.583 0.9129
#> det.cov.4.1  0.8623 1.2024 1.3944 1.599 2.0311
#>