Function for Fitting a Latent Factor Joint Species Distribution Model

Function for fitting a joint species distribution model with species correlations. This model does not explicitly account for imperfect detection (see lfMsPGOcc()). We use Polya-gamma latent variables and a factor modeling approach.

Usage

lfJSDM(formula, data, inits, priors, n.factors, 
       n.samples, n.omp.threads = 1, verbose = TRUE, n.report = 100, 
       n.burn = round(.10 * n.samples), n.thin = 1, n.chains = 1,
       k.fold, k.fold.threads = 1, k.fold.seed, k.fold.only = FALSE, ...)

Arguments

formula

a symbolic description of the model to be fit for the model using R's model syntax. Only right-hand side of formula is specified. See example below. Random intercepts are allowed using lme4 syntax (Bates et al. 2015).

data

a list containing data necessary for model fitting. Valid tags are y, covs, and coords. y is a two-dimensional array with first dimension equal to the number of species and second dimension equal to the number of sites. Note how this differs from other spOccupancy functions in that y does not have any replicate surveys. This is because lfJSDM does not account for imperfect detection. covs is a matrix or data frame containing the variables used in the model, with \(J\) rows for each column (variable). coords is a matrix with \(J\) rows and 2 columns consisting of the spatial coordinates of each site in the data. Note that spOccupancy assumes coordinates are specified in a projected coordinate system.

inits

a list with each tag corresponding to a parameter name. Valid tags are beta.comm, beta, tau.sq.beta, sigma.sq.psi, lambda. The value portion of each tag is the parameter's initial value. See priors description for definition of each parameter name. Additionally, the tag fix can be set to TRUE to fix the starting values across all chains. If fix is not specified (the default), starting values are varied randomly across chains.

priors

a list with each tag corresponding to a parameter name. Valid tags are beta.comm.normal, tau.sq.beta.ig, and sigma.sq.psi.ig. Community-level (beta.comm) regression coefficients are assumed to follow a normal distribution. The hyperparameters of the normal distribution are passed as a list of length two with the first and second elements corresponding to the mean and variance of the normal distribution, which are each specified as vectors of length equal to the number of coefficients to be estimated or of length one if priors are the same for all coefficients. If not specified, prior means are set to 0 and prior variances set to 2.72. Community-level variance parameters (tau.sq.beta) are assumed to follow an inverse Gamma distribution. The hyperparameters of the inverse gamma distribution are passed as a list of length two with the first and second elements corresponding to the shape and scale parameters, which are each specified as vectors of length equal to the number of coefficients to be estimated or a single value if all parameters are assigned the same prior. If not specified, prior shape and scale parameters are set to 0.1. The factor model fits n.factors independent latent factors. The priors for the factor loadings matrix lambda are fixed following standard approaches to ensure parameter identifiability. The upper triangular elements of the N x n.factors matrix are fixed at 0 and the diagonal elements are fixed at 1. The lower triangular elements are assigned a standard normal prior (i.e., mean 0 and variance 1). sigma.sq.psi is the random effect variance for any random effects, and is assumed to follow an inverse Gamma distribution. The hyperparameters of the inverse-Gamma distribution are passed as a list of length two with first and second elements corresponding to the shape and scale parameters, respectively, which are each specified as vectors of length equal to the number of random intercepts or of length one if priors are the same for all random effect variances.

n.factors

the number of factors to use in the latent factor model approach. Typically, the number of factors is set to be small (e.g., 4-5) relative to the total number of species in the community, which will lead to substantial decreases in computation time. However, the value can be anywhere between 1 and N (the number of species in the community).

n.samples

the number of posterior samples to collect in each chain.

n.omp.threads

a positive integer indicating the number of threads to use for SMP parallel processing. The package must be compiled for OpenMP support. For most Intel-based machines, we recommend setting n.omp.threads up to the number of hypterthreaded cores. Note, n.omp.threads > 1 might not work on some systems.

verbose

if TRUE, messages about data preparation, model specification, and progress of the sampler are printed to the screen. Otherwise, no messages are printed.

n.report

the interval to report MCMC progress.

n.burn

the number of samples out of the total n.samples to discard as burn-in for each chain. By default, the first 10% of samples is discarded.

n.thin

the thinning interval for collection of MCMC samples. The thinning occurs after the n.burn samples are discarded. Default value is set to 1.

n.chains

the number of chains to run in sequence.

k.fold

specifies the number of k folds for cross-validation. If not specified as an argument, then cross-validation is not performed and k.fold.threads and k.fold.seed are ignored. In k-fold cross-validation, the data specified in data is randomly partitioned into k equal sized subsamples. Of the k subsamples, k - 1 subsamples are used to fit the model and the remaining k samples are used for prediction. The cross-validation process is repeated k times (the folds). As a scoring rule, we use the model deviance as described in Hooten and Hobbs (2015). Cross-validation is performed after the full model is fit using all the data. Cross-validation results are reported in the k.fold.deviance object in the return list.

k.fold.threads

number of threads to use for cross-validation. If k.fold.threads > 1 parallel processing is accomplished using the foreach and doParallel packages. Ignored if k.fold is not specified.

k.fold.seed

seed used to split data set into k.fold parts for k-fold cross-validation. Ignored if k.fold is not specified.

k.fold.only

a logical value indicating whether to only perform cross-validation (TRUE) or perform cross-validation after fitting the full model (FALSE). Default value is FALSE.

...

currently no additional arguments

Note

Some of the underlying code used for generating random numbers from the Polya-Gamma distribution is taken from the pgdraw package written by Daniel F. Schmidt and Enes Makalic. Their code implements Algorithm 6 in PhD thesis of Jesse Bennett Windle (2013) https://repositories.lib.utexas.edu/handle/2152/21842.

References

Polson, N.G., J.G. Scott, and J. Windle. (2013) Bayesian Inference for Logistic Models Using Polya-Gamma Latent Variables. Journal of the American Statistical Association, 108:1339-1349.

Bates, Douglas, Martin Maechler, Ben Bolker, Steve Walker (2015). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67(1), 1-48. doi:10.18637/jss.v067.i01 .

Hooten, M. B., and Hobbs, N. T. (2015). A guide to Bayesian model selection for ecologists. Ecological monographs, 85(1), 3-28.

Author

Jeffrey W. Doser doserjef@msu.edu,
Andrew O. Finley finleya@msu.edu

Value

An object of class lfJSDM that is a list comprised of:

beta.comm.samples

a coda object of posterior samples for the community level occurrence regression coefficients.

tau.sq.beta.samples

a coda object of posterior samples for the occurrence community variance parameters.

beta.samples

a coda object of posterior samples for the species level occurrence regression coefficients.

lambda.samples

a coda object of posterior samples for the latent factor loadings.

psi.samples

a three-dimensional array of posterior samples for the latent probability of occurrence/detection values for each species.

sigma.sq.psi.samples

a coda object of posterior samples for variances of random intercepts included in the occurrence portion of the model. Only included if random intercepts are specified in occ.formula.

w.samples

a three-dimensional array of posterior samples for the latent effects for each latent factor.

beta.star.samples

a coda object of posterior samples for the occurrence random effects. Only included if random intercepts are specified in occ.formula.

like.samples

a three-dimensional array of posterior samples for the likelihood value associated with each site and species. Used for calculating WAIC.

rhat

a list of Gelman-Rubin diagnostic values for some of the model parameters.

ESS

a list of effective sample sizes for some of the model parameters.

run.time

MCMC sampler execution time reported using proc.time().

k.fold.deviance

vector of scoring rules (deviance) from k-fold cross-validation. A separate value is reported for each species. Only included if k.fold is specified in function call.

The return object will include additional objects used for subsequent prediction and/or model fit evaluation. Note that detection probability estimated values are not included in the model object, but can be extracted using fitted().

Examples

set.seed(400)
J.x <- 10
J.y <- 10
J <- J.x * J.y
n.rep <- rep(1, J)
N <- 10
# Community-level covariate effects
# Occurrence
beta.mean <- c(0.2, 0.6, 1.5)
p.occ <- length(beta.mean)
tau.sq.beta <- c(0.6, 1.2, 1.7)
# Detection
# Fix this to be constant and really close to 1. 
alpha.mean <- c(9)
tau.sq.alpha <- c(0.05)
p.det <- length(alpha.mean)
# Random effects
# Include a single random effect
psi.RE <- list(levels = c(20), 
               sigma.sq.psi = c(2))
p.RE <- list()
# Draw species-level effects from community means.
beta <- matrix(NA, nrow = N, ncol = p.occ)
alpha <- matrix(NA, nrow = N, ncol = p.det)
for (i in 1:p.occ) {
  beta[, i] <- rnorm(N, beta.mean[i], sqrt(tau.sq.beta[i]))
}
for (i in 1:p.det) {
  alpha[, i] <- rnorm(N, alpha.mean[i], sqrt(tau.sq.alpha[i]))
}
alpha.true <- alpha
# Factor model
factor.model <- TRUE
n.factors <- 4

dat <- simMsOcc(J.x = J.x, J.y = J.y, n.rep = n.rep, N = N, beta = beta, alpha = alpha,
                psi.RE = psi.RE, p.RE = p.RE, sp = FALSE,
                factor.model = TRUE, n.factors = 4)

X <- dat$X
y <- dat$y
X.re <- dat$X.re
coords <- dat$coords
occ.covs <- cbind(X, X.re)
colnames(occ.covs) <- c('int', 'occ.cov.1', 'occ.cov.2', 'occ.re.1')
data.list <- list(y = y[, , 1], 
                  covs = occ.covs, 
                  coords = coords) 
# Priors
prior.list <- list(beta.comm.normal = list(mean = 0, var = 2.72),
                   tau.sq.beta.ig = list(a = 0.1, b = 0.1)) 
inits.list <- list(beta.comm = 0, beta = 0, tau.sq.beta = 1) 
out <- lfJSDM(formula = ~ occ.cov.1 + occ.cov.2 + (1 | occ.re.1), 
              data = data.list, 
              inits = inits.list, 
              priors = prior.list, 
              n.factors = 4, 
              n.samples = 1000,
              n.report = 500, 
              n.burn = 500,
              n.thin = 2,
              n.chains = 1) 
#> ----------------------------------------
#> 	Preparing to run the model
#> ----------------------------------------
#> No prior specified for sigma.sq.psi.ig.
#> Setting prior shape to 0.1 and prior scale to 0.1
#> lambda is not specified in initial values.
#> Setting initial values of the lower triangle to random values from a standard normal
#> sigma.sq.psi is not specified in initial values.
#> Setting initial values to random values between 0.5 and 10
#> ----------------------------------------
#> 	Model description
#> ----------------------------------------
#> Latent Factor JSDM with Polya-Gamma latent
#> variable fit with 100 sites and 10 species.
#> 
#> Samples per Chain: 1000 
#> Burn-in: 500 
#> Thinning Rate: 2 
#> Number of Chains: 1 
#> Total Posterior Samples: 250 
#> 
#> Using 4 latent factors.
#> 
#> Source compiled with OpenMP support and model fit using 1 thread(s).
#> 
#> ----------------------------------------
#> 	Chain 1
#> ----------------------------------------
#> Sampling ... 
#> Sampled: 500 of 1000, 50.00%
#> -------------------------------------------------
#> Sampled: 1000 of 1000, 100.00%
summary(out)
#> 
#> Call:
#> lfJSDM(formula = ~occ.cov.1 + occ.cov.2 + (1 | occ.re.1), data = data.list, 
#>     inits = inits.list, priors = prior.list, n.factors = 4, n.samples = 1000, 
#>     n.report = 500, n.burn = 500, n.thin = 2, n.chains = 1)
#> 
#> Samples per Chain: 1000
#> Burn-in: 500
#> Thinning Rate: 2
#> Number of Chains: 1
#> Total Posterior Samples: 250
#> Run Time (min): 0.0181
#> 
#> ----------------------------------------
#> 	Community Level
#> ----------------------------------------
#> Means (logit scale): 
#>                Mean     SD    2.5%     50%  97.5% Rhat ESS
#> (Intercept)  0.3856 0.2885 -0.1799  0.3652 0.9046   NA 130
#> occ.cov.1   -0.0161 0.2417 -0.4897 -0.0090 0.4239   NA 149
#> occ.cov.2    1.8966 0.5645  0.7120  1.9592 2.9398   NA 250
#> 
#> Variances (logit scale): 
#>               Mean     SD   2.5%    50%  97.5% Rhat ESS
#> (Intercept) 0.4819 0.4312 0.0959 0.3652 1.7115   NA  37
#> occ.cov.1   0.4279 0.3311 0.0788 0.3424 1.4497   NA  94
#> occ.cov.2   3.1386 1.7818 1.0629 2.6556 7.4200   NA 178
#> 
#> Random Effect Variances (logit scale): 
#>            Mean     SD  2.5%    50%  97.5% Rhat ESS
#> occ.re.1 2.5587 0.8536 1.329 2.4116 4.7669   NA  12
#> 
#> ----------------------------------------
#> 	Species Level
#> ----------------------------------------
#> Estimates (logit scale): 
#>                     Mean     SD    2.5%     50%   97.5% Rhat ESS
#> (Intercept)-sp1   0.0058 0.4211 -0.8075  0.0080  0.8035   NA 105
#> (Intercept)-sp2   0.3123 0.4212 -0.5193  0.3193  1.1159   NA  89
#> (Intercept)-sp3   0.7817 0.4200  0.0141  0.7752  1.5493   NA  80
#> (Intercept)-sp4   0.1666 0.4243 -0.6005  0.1492  0.9344   NA  82
#> (Intercept)-sp5   0.3722 0.3871 -0.3064  0.3683  1.1603   NA  82
#> (Intercept)-sp6  -0.4095 0.5087 -1.6071 -0.3568  0.4230   NA  52
#> (Intercept)-sp7   0.8326 0.4684 -0.0903  0.8172  1.7304   NA  76
#> (Intercept)-sp8   0.4618 0.3897 -0.2672  0.4670  1.2079   NA 103
#> (Intercept)-sp9   1.1546 0.5000  0.2501  1.1171  2.2212   NA  57
#> (Intercept)-sp10  0.2765 0.3998 -0.5082  0.2719  1.0559   NA  81
#> occ.cov.1-sp1    -0.5906 0.3968 -1.4261 -0.5706  0.1490   NA  96
#> occ.cov.1-sp2     0.6571 0.3395  0.0051  0.6383  1.3031   NA 108
#> occ.cov.1-sp3    -0.5230 0.3378 -1.2858 -0.5254  0.1252   NA  46
#> occ.cov.1-sp4    -0.0296 0.3465 -0.6811  0.0192  0.6286   NA 126
#> occ.cov.1-sp5    -0.2289 0.3516 -0.9708 -0.2096  0.5018   NA 129
#> occ.cov.1-sp6     0.4564 0.3600 -0.2633  0.4459  1.2761   NA  82
#> occ.cov.1-sp7     0.3578 0.3758 -0.2265  0.3041  1.1700   NA  82
#> occ.cov.1-sp8     0.0565 0.3313 -0.6476  0.0800  0.7322   NA 141
#> occ.cov.1-sp9     0.2791 0.3475 -0.3502  0.2578  1.0533   NA 128
#> occ.cov.1-sp10   -0.5631 0.4334 -1.4986 -0.5194  0.1507   NA 122
#> occ.cov.2-sp1     3.8952 0.6593  2.7759  3.8575  5.1451   NA  65
#> occ.cov.2-sp2    -1.2642 0.4030 -2.0969 -1.2449 -0.5434   NA 117
#> occ.cov.2-sp3     1.3880 0.3841  0.7131  1.3544  2.1726   NA  50
#> occ.cov.2-sp4     2.7743 0.5661  1.7639  2.7499  3.8817   NA  52
#> occ.cov.2-sp5     2.9446 0.6475  1.8759  2.9277  4.2666   NA  45
#> occ.cov.2-sp6     3.0263 0.7554  1.8829  2.8990  4.7376   NA  45
#> occ.cov.2-sp7     3.6107 0.7544  2.3841  3.5449  5.1560   NA  31
#> occ.cov.2-sp8     1.6771 0.5217  0.7526  1.6913  2.7278   NA  63
#> occ.cov.2-sp9     2.2531 0.5341  1.3009  2.2146  3.3277   NA  60
#> occ.cov.2-sp10    0.8913 0.4481  0.0988  0.8303  1.8067   NA  64