Function for prediction at new locations for latent factor joint species distribution models

The function predict collects posterior predictive samples for a set of new locations given an object of class `lfJSDM`.

# S3 method for lfJSDM
predict(object, X.0, coords.0, 
        ignore.RE = FALSE, ...)

Arguments

object

an object of class lfJSDM

X.0

the design matrix of covariates at the prediction locations. This should include a column of 1s for the intercept if an intercept is included in the model. If random effects are included in the model, the levels of the random effects at the new locations should be included as a column in the design matrix. The ordering of the levels should match the ordering used to fit the data in lfJSDM. Columns should correspond to the order of how covariates were specified in the formula argument of lfJSDM. Column names of the random effects must match the name of the random effects, if specified in the formula argument of lfJSDM.

coords.0

the spatial coordinates corresponding to X.0. Note that spOccupancy assumes coordinates are specified in a projected coordinate system.

ignore.RE

a logical value indicating whether to include unstructured random effects for prediction. If TRUE, random effects will be ignored and prediction will only use the fixed effects. If FALSE, random effects will be included in the prediction for both observed and unobserved levels of the random effect.

...

currently no additional arguments

Note

When ignore.RE = FALSE, both sampled levels and non-sampled levels of random effects are supported for prediction. For sampled levels, the posterior distribution for the random intercept corresponding to that level of the random effect will be used in the prediction. For non-sampled levels, random values are drawn from a normal distribution using the posterior samples of the random effect variance, which results in fully propagated uncertainty in predictions with models that incorporate random effects.

Author

Jeffrey W. Doser doserjef@msu.edu,
Andrew O. Finley finleya@msu.edu

Value

A list object of class predict.lfJSDM that consists of:

psi.0.samples

a three-dimensional array of posterior predictive samples for the latent occurrence probability values.

z.0.samples

a three-dimensional array of posterior predictive samples for the latent occurrence values.

w.0.samples

a three-dimensional array of posterior predictive samples for the latent factors.

The return object will include additional objects used for standard extractor functions.

Examples

set.seed(400)
J.x <- 8
J.y <- 8
J <- J.x * J.y
n.rep<- sample(2:4, size = J, replace = TRUE)
N <- 6
# Community-level covariate effects
# Occurrence
beta.mean <- c(0.2, 0.5)
p.occ <- length(beta.mean)
tau.sq.beta <- c(0.6, 0.3)
# Detection
alpha.mean <- c(0.5, 0.2, -0.1)
tau.sq.alpha <- c(0.2, 0.3, 1)
p.det <- length(alpha.mean)
# Draw species-level effects from community means.
beta <- matrix(NA, nrow = N, ncol = p.occ)
alpha <- matrix(NA, nrow = N, ncol = p.det)
for (i in 1:p.occ) {
  beta[, i] <- rnorm(N, beta.mean[i], sqrt(tau.sq.beta[i]))
}
for (i in 1:p.det) {
  alpha[, i] <- rnorm(N, alpha.mean[i], sqrt(tau.sq.alpha[i]))
}

n.factors <- 3
dat <- simMsOcc(J.x = J.x, J.y = J.y, n.rep = n.rep, N = N, beta = beta, alpha = alpha,
                sp = FALSE, factor.model = TRUE, n.factors = n.factors)
n.samples <- 5000
# Split into fitting and prediction data set
pred.indx <- sample(1:J, round(J * .25), replace = FALSE)
# Summarize the multiple replicates into a single value for use in a JSDM
y <- apply(dat$y[, -pred.indx, ], c(1, 2), max, na.rm = TRUE)
# Covariates
X <- dat$X[-pred.indx, ]
# Spatial coordinates
coords <- dat$coords[-pred.indx, ]
# Prediction values
X.0 <- dat$X[pred.indx, ]
psi.0 <- dat$psi[, pred.indx]
coords.0 <- dat$coords[pred.indx, ]
# Package all data into a list
covs <- X[, 2, drop = FALSE]
colnames(covs) <- c('occ.cov')
data.list <- list(y = y, 
                  covs = covs,
                  coords = coords)

# Occupancy initial values
prior.list <- list(beta.comm.normal = list(mean = 0, var = 2.72), 
                   tau.sq.beta.ig = list(a = 0.1, b = 0.1))
# Initial values
lambda.inits <- matrix(0, N, n.factors)
diag(lambda.inits) <- 1
lambda.inits[lower.tri(lambda.inits)] <- rnorm(sum(lower.tri(lambda.inits)))
inits.list <- list(alpha.comm = 0, 
                   beta.comm = 0, 
                   beta = 0, 
                   tau.sq.beta = 1, 
                   lambda = lambda.inits)

out <- lfJSDM(formula = ~ occ.cov, 
              data = data.list, 
              inits = inits.list, 
              n.samples = n.samples, 
              n.factors = 3, 
              priors = prior.list, 
              n.omp.threads = 1, 
              verbose = TRUE, 
              n.report = 1000, 
              n.burn = 4000)
#> ----------------------------------------
#> 	Preparing to run the model
#> ----------------------------------------
#> ----------------------------------------
#> 	Model description
#> ----------------------------------------
#> Latent Factor JSDM with Polya-Gamma latent
#> variable fit with 48 sites and 6 species.
#> 
#> Samples per Chain: 5000 
#> Burn-in: 4000 
#> Thinning Rate: 1 
#> Number of Chains: 1 
#> Total Posterior Samples: 1000 
#> 
#> Using 3 latent factors.
#> 
#> Source compiled with OpenMP support and model fit using 1 thread(s).
#> 
#> ----------------------------------------
#> 	Chain 1
#> ----------------------------------------
#> Sampling ... 
#> Sampled: 1000 of 5000, 20.00%
#> -------------------------------------------------
#> Sampled: 2000 of 5000, 40.00%
#> -------------------------------------------------
#> Sampled: 3000 of 5000, 60.00%
#> -------------------------------------------------
#> Sampled: 4000 of 5000, 80.00%
#> -------------------------------------------------
#> Sampled: 5000 of 5000, 100.00%

summary(out)
#> 
#> Call:
#> lfJSDM(formula = ~occ.cov, data = data.list, inits = inits.list, 
#>     priors = prior.list, n.factors = 3, n.samples = n.samples, 
#>     n.omp.threads = 1, verbose = TRUE, n.report = 1000, n.burn = 4000)
#> 
#> Samples per Chain: 5000
#> Burn-in: 4000
#> Thinning Rate: 1
#> Number of Chains: 1
#> Total Posterior Samples: 1000
#> Run Time (min): 0.0195
#> 
#> ----------------------------------------
#> 	Community Level
#> ----------------------------------------
#> Means (logit scale): 
#>                Mean     SD    2.5%     50%  97.5% Rhat ESS
#> (Intercept) -0.2264 0.3564 -0.9776 -0.2279 0.4647   NA 726
#> occ.cov      0.1344 0.3795 -0.6306  0.1346 0.9220   NA 657
#> 
#> Variances (logit scale): 
#>               Mean     SD   2.5%    50%  97.5% Rhat ESS
#> (Intercept) 0.7147 0.7847 0.0968 0.4753 2.7119   NA 449
#> occ.cov     0.7405 0.9059 0.0792 0.4687 2.8696   NA 481
#> 
#> ----------------------------------------
#> 	Species Level
#> ----------------------------------------
#> Estimates (logit scale): 
#>                    Mean     SD    2.5%     50%   97.5% Rhat ESS
#> (Intercept)-sp1  0.6551 0.3896 -0.0823  0.6381  1.4265   NA 370
#> (Intercept)-sp2 -0.4163 0.3733 -1.1879 -0.4010  0.2674   NA 530
#> (Intercept)-sp3 -0.2525 0.3980 -1.0479 -0.2483  0.5224   NA 412
#> (Intercept)-sp4  0.0946 0.3863 -0.6988  0.0969  0.9133   NA 277
#> (Intercept)-sp5 -0.6274 0.4024 -1.4841 -0.6117  0.1188   NA 414
#> (Intercept)-sp6 -0.8831 0.4106 -1.7356 -0.8600 -0.1373   NA 328
#> occ.cov-sp1      0.7989 0.4010  0.1021  0.7775  1.6235   NA 301
#> occ.cov-sp2     -0.1790 0.3520 -0.8811 -0.1779  0.4601   NA 327
#> occ.cov-sp3      0.2222 0.3667 -0.4661  0.2098  0.9315   NA 382
#> occ.cov-sp4      0.7002 0.4318 -0.0467  0.6630  1.6013   NA 297
#> occ.cov-sp5     -0.1622 0.3586 -0.8989 -0.1483  0.5125   NA 395
#> occ.cov-sp6     -0.6380 0.4293 -1.6449 -0.6145  0.0877   NA 277

# Predict at new locations ------------------------------------------------
out.pred <- predict(out, X.0, coords.0)