Function for prediction at new locations for multi-species spatial occupancy models

The function predict collects posterior predictive samples for a set of new locations given an object of class `spMsPGOcc`. Prediction is possible for both the latent occupancy state as well as detection.

Usage

# S3 method for spMsPGOcc
predict(object, X.0, coords.0, n.omp.threads = 1, verbose = TRUE, 
                            n.report = 100, ignore.RE = FALSE, type = 'occupancy', ...)

Arguments

object

an object of class spMsPGOcc

X.0

the design matrix of covariates at the prediction locations. This should include a column of 1s for the intercept if an intercept is included in the model. If random effects are included in the occupancy (or detection if type = 'detection') portion of the model, the levels of the random effects at the new locations should be included as a column in the design matrix. The ordering of the levels should match the ordering used to fit the data in spMsPGOcc. Columns should correspond to the order of how covariates were specified in the corresponding formula argument of spMsPGOcc. Column names of the random effects must match the name of the random effects, if specified in the corresponding formula argument of spMsPGOcc.

coords.0

the spatial coordinates corresponding to X.0. Note that spOccupancy assumes coordinates are specified in a projected coordinate system.

n.omp.threads

a positive integer indicating the number of threads to use for SMP parallel processing. The package must be compiled for OpenMP support. For most Intel-based machines, we recommend setting n.omp.threads up to the number of hyperthreaded cores. Note, n.omp.threads > 1 might not work on some systems.

verbose

if TRUE, model specification and progress of the sampler is printed to the screen. Otherwise, nothing is printed to the screen.

n.report

the interval to report sampling progress.

ignore.RE

a logical value indicating whether to include unstructured random effects for prediction. If TRUE, unstructured random effects will be ignored and prediction will only use the fixed effects and the spatial random effects. If FALSE, random effects will be included in the prediction for both observed and unobserved levels of the unstructured random effects.

type

a quoted keyword indicating what type of prediction to produce. Valid keywords are 'occupancy' to predict latent occupancy probability and latent occupancy values (this is the default), or 'detection' to predict detection probability given new values of detection covariates.

...

currently no additional arguments

Note

When ignore.RE = FALSE, both sampled levels and non-sampled levels of random effects are supported for prediction. For sampled levels, the posterior distribution for the random intercept corresponding to that level of the random effect will be used in the prediction. For non-sampled levels, random values are drawn from a normal distribution using the posterior samples of the random effect variance, which results in fully propagated uncertainty in predictions with models that incorporate random effects.

Author

Jeffrey W. Doser doserjef@msu.edu,
Andrew O. Finley finleya@msu.edu

Value

An list object of class predict.spMsPGOcc. When type = 'occupancy', the list consists of:

psi.0.samples

a three-dimensional array of posterior predictive samples for the latent occurrence probability values.

z.0.samples

a three-dimensional array of posterior predictive samples for the latent occurrence values.

w.0.samples

a three-dimensional array of posterior predictive samples for the latent spatial random effects.

run.time

execution time reported using proc.time().

When type = 'detection', the list consists of:

p.0.samples

a three-dimensional array of posterior predictive samples for the detection probability values.

run.time

execution time reported using proc.time().

The return object will include additional objects used for standard extractor functions.

Examples

set.seed(400)

# Simulate Data -----------------------------------------------------------
J.x <- 7
J.y <- 7
J <- J.x * J.y
n.rep <- sample(2:4, size = J, replace = TRUE)
N <- 5
# Community-level covariate effects
# Occurrence
beta.mean <- c(0.2, -0.15)
p.occ <- length(beta.mean)
tau.sq.beta <- c(0.6, 0.3)
# Detection
alpha.mean <- c(0.5, 0.2, -.2)
tau.sq.alpha <- c(0.2, 0.3, 0.8)
p.det <- length(alpha.mean)
# Draw species-level effects from community means.
beta <- matrix(NA, nrow = N, ncol = p.occ)
alpha <- matrix(NA, nrow = N, ncol = p.det)
for (i in 1:p.occ) {
  beta[, i] <- rnorm(N, beta.mean[i], sqrt(tau.sq.beta[i]))
}
for (i in 1:p.det) {
  alpha[, i] <- rnorm(N, alpha.mean[i], sqrt(tau.sq.alpha[i]))
}
phi <- runif(N, 3/1, 3/.4)
sigma.sq <- runif(N, 0.3, 3)
sp <- TRUE

dat <- simMsOcc(J.x = J.x, J.y = J.y, n.rep = n.rep, N = N, beta = beta, alpha = alpha,
    phi = phi, sigma.sq = sigma.sq, sp = TRUE, cov.model = 'exponential')

# Number of batches
n.batch <- 30
# Batch length
batch.length <- 25
n.samples <- n.batch * batch.length

# Split into fitting and prediction data set
pred.indx <- sample(1:J, round(J * .25), replace = FALSE)
y <- dat$y[, -pred.indx, ]
# Occupancy covariates
X <- dat$X[-pred.indx, ]
# Coordinates
coords <- as.matrix(dat$coords[-pred.indx, ])
# Detection covariates
X.p <- dat$X.p[-pred.indx, , ]
# Prediction values
X.0 <- dat$X[pred.indx, ]
coords.0 <- as.matrix(dat$coords[pred.indx, ])
psi.0 <- dat$psi[, pred.indx]

# Package all data into a list
occ.covs <- X[, 2, drop = FALSE]
colnames(occ.covs) <- c('occ.cov')
det.covs <- list(det.cov.1 = X.p[, , 2], 
     det.cov.2 = X.p[, , 3]
     )
data.list <- list(y = y, 
      occ.covs = occ.covs,
      det.covs = det.covs, 
      coords = coords)

# Priors 
prior.list <- list(beta.comm.normal = list(mean = 0, var = 2.72), 
       alpha.comm.normal = list(mean = 0, var = 2.72), 
       tau.sq.beta.ig = list(a = 0.1, b = 0.1), 
       tau.sq.alpha.ig = list(a = 0.1, b = 0.1),
       phi.unif = list(a = 3/1, b = 3/.1), 
       sigma.sq.ig = list(a = 2, b = 2)) 
# Starting values
inits.list <- list(alpha.comm = 0, 
          beta.comm = 0, 
          beta = 0, 
          alpha = 0,
          tau.sq.beta = 1, 
          tau.sq.alpha = 1, 
          phi = 3 / .5, 
          sigma.sq = 2,
          w = matrix(0, nrow = N, ncol = nrow(X)),
          z = apply(y, c(1, 2), max, na.rm = TRUE))
# Tuning
tuning.list <- list(phi = 1) 

out <- spMsPGOcc(occ.formula = ~ occ.cov, 
     det.formula = ~ det.cov.1 + det.cov.2, 
     data = data.list,
     inits = inits.list, 
     n.batch = n.batch, 
     batch.length = batch.length, 
     accept.rate = 0.43, 
           priors = prior.list, 
     cov.model = "exponential", 
     tuning = tuning.list, 
           n.omp.threads = 1, 
           verbose = TRUE, 
     NNGP = TRUE, 
     n.neighbors = 5, 
     search.type = 'cb', 
           n.report = 10, 
     n.burn = 500, 
     n.thin = 1)
#> ----------------------------------------
#> 	Preparing to run the model
#> ----------------------------------------
#> ----------------------------------------
#> 	Building the neighbor list
#> ----------------------------------------
#> ----------------------------------------
#> Building the neighbors of neighbors list
#> ----------------------------------------
#> ----------------------------------------
#> 	Model description
#> ----------------------------------------
#> NNGP Multispecies Occupancy Model with Polya-Gamma latent
#> variable fit with 37 sites and 5 species.
#> 
#> Samples per chain: 750 (30 batches of length 25)
#> Burn-in: 500 
#> Thinning Rate: 1 
#> Number of Chains: 1 
#> Total Posterior Samples: 250 
#> 
#> Using the exponential spatial correlation model.
#> 
#> Using 5 nearest neighbors.
#> 
#> Source compiled with OpenMP support and model fit using 1 thread(s).
#> 
#> Adaptive Metropolis with target acceptance rate: 43.0
#> ----------------------------------------
#> 	Chain 1
#> ----------------------------------------
#> Sampling ... 
#> Batch: 10 of 30, 33.33%
#> 	Species		Parameter	Acceptance	Tuning
#> 	1		phi		64.0		1.11628
#> 	2		phi		76.0		1.11628
#> 	3		phi		80.0		1.11628
#> 	4		phi		80.0		1.11628
#> 	5		phi		68.0		1.11628
#> -------------------------------------------------
#> Batch: 20 of 30, 66.67%
#> 	Species		Parameter	Acceptance	Tuning
#> 	1		phi		68.0		1.23368
#> 	2		phi		84.0		1.23368
#> 	3		phi		52.0		1.23368
#> 	4		phi		80.0		1.23368
#> 	5		phi		80.0		1.23368
#> -------------------------------------------------
#> Batch: 30 of 30, 100.00%

summary(out, level = 'both')
#> 
#> Call:
#> spMsPGOcc(occ.formula = ~occ.cov, det.formula = ~det.cov.1 + 
#>     det.cov.2, data = data.list, inits = inits.list, priors = prior.list, 
#>     tuning = tuning.list, cov.model = "exponential", NNGP = TRUE, 
#>     n.neighbors = 5, search.type = "cb", n.batch = n.batch, batch.length = batch.length, 
#>     accept.rate = 0.43, n.omp.threads = 1, verbose = TRUE, n.report = 10, 
#>     n.burn = 500, n.thin = 1)
#> 
#> Samples per Chain: 750
#> Burn-in: 500
#> Thinning Rate: 1
#> Number of Chains: 1
#> Total Posterior Samples: 250
#> Run Time (min): 0.02
#> 
#> ----------------------------------------
#> 	Community Level
#> ----------------------------------------
#> Occurrence Means (logit scale): 
#>               Mean     SD    2.5%    50%  97.5% Rhat ESS
#> (Intercept) 0.1502 0.4968 -0.7868 0.1471 1.1698   NA  76
#> occ.cov     0.0015 0.4914 -0.9922 0.0010 0.9001   NA 162
#> 
#> Occurrence Variances (logit scale): 
#>               Mean     SD   2.5%    50%  97.5% Rhat ESS
#> (Intercept) 0.7201 1.1839 0.0523 0.3720 3.3979   NA  27
#> occ.cov     1.1692 1.8236 0.0713 0.5796 7.2585   NA 142
#> 
#> Detection Means (logit scale): 
#>                Mean     SD    2.5%     50%  97.5% Rhat ESS
#> (Intercept)  0.4424 0.4641 -0.5687  0.4470 1.3007   NA 250
#> det.cov.1    0.2085 0.5944 -1.0772  0.2383 1.2827   NA 284
#> det.cov.2   -0.8330 0.6754 -2.0404 -0.8223 0.5206   NA 250
#> 
#> Detection Variances (logit scale): 
#>               Mean     SD   2.5%    50%   97.5% Rhat ESS
#> (Intercept) 1.1829 1.6181 0.1498 0.6000  6.4956   NA 113
#> det.cov.1   1.3905 1.5723 0.1579 0.8197  6.5105   NA  72
#> det.cov.2   2.9168 6.9862 0.1540 1.2750 16.4065   NA 143
#> 
#> ----------------------------------------
#> 	Species Level
#> ----------------------------------------
#> Occurrence (logit scale): 
#>                    Mean     SD    2.5%     50%  97.5% Rhat ESS
#> (Intercept)-sp1  0.1705 0.4725 -0.6416  0.1824 1.0501   NA  42
#> (Intercept)-sp2  0.7750 0.4912  0.0299  0.7514 1.8524   NA  58
#> (Intercept)-sp3 -0.3799 0.5715 -1.5790 -0.3278 0.6004   NA  38
#> (Intercept)-sp4  0.2861 0.5304 -0.5145  0.2464 1.6364   NA  55
#> (Intercept)-sp5 -0.0494 0.4462 -0.8993 -0.0213 0.7472   NA  67
#> occ.cov-sp1     -0.3791 0.4276 -1.2436 -0.3857 0.4727   NA 109
#> occ.cov-sp2      0.1140 0.4366 -0.8054  0.1382 0.9200   NA  85
#> occ.cov-sp3     -0.7358 0.6016 -2.0397 -0.6806 0.2713   NA  53
#> occ.cov-sp4      0.3419 0.5209 -0.6432  0.3066 1.3924   NA  37
#> occ.cov-sp5      0.6949 0.6888 -0.3877  0.5845 2.2906   NA  33
#> 
#> Detection (logit scale): 
#>                    Mean     SD    2.5%     50%   97.5% Rhat ESS
#> (Intercept)-sp1  1.4782 0.4319  0.7387  1.4712  2.3217   NA  59
#> (Intercept)-sp2  1.0440 0.2973  0.5109  1.0286  1.6560   NA  99
#> (Intercept)-sp3 -0.0786 0.5794 -1.1580 -0.0749  1.0515   NA  64
#> (Intercept)-sp4 -0.1595 0.3598 -0.8475 -0.1584  0.4787   NA  48
#> (Intercept)-sp5  0.2127 0.3466 -0.4585  0.2152  0.8647   NA 115
#> det.cov.1-sp1    0.6951 0.3761 -0.0018  0.6577  1.4108   NA  64
#> det.cov.1-sp2   -0.4873 0.2859 -1.0588 -0.4875  0.0677   NA 216
#> det.cov.1-sp3    1.4250 0.7240  0.4015  1.3715  2.9166   NA  46
#> det.cov.1-sp4   -0.3892 0.3035 -1.0714 -0.3646  0.1821   NA  92
#> det.cov.1-sp5    0.2397 0.3282 -0.3950  0.2344  0.9884   NA 145
#> det.cov.2-sp1   -0.0960 0.3024 -0.6983 -0.0621  0.4125   NA  65
#> det.cov.2-sp2   -0.6448 0.2963 -1.2741 -0.6285 -0.1040   NA 125
#> det.cov.2-sp3   -2.8161 1.2517 -5.7716 -2.6072 -1.0448   NA  20
#> det.cov.2-sp4   -0.3190 0.2547 -0.8455 -0.3229  0.1931   NA 111
#> det.cov.2-sp5   -1.2080 0.4038 -2.1116 -1.1799 -0.5214   NA  92
#> 
#> Spatial Covariance: 
#>                 Mean     SD   2.5%     50%   97.5% Rhat ESS
#> sigma.sq-sp1  2.3331 2.5098 0.3751  1.5425 10.8606   NA  30
#> sigma.sq-sp2  1.7216 1.1344 0.3795  1.5075  4.3097   NA  20
#> sigma.sq-sp3  2.2486 1.6933 0.5808  1.8402  7.6037   NA  27
#> sigma.sq-sp4  1.6693 1.3469 0.3570  1.2919  5.1814   NA  15
#> sigma.sq-sp5  6.0948 8.9513 0.3599  1.6360 32.0934   NA   5
#> phi-sp1      14.7813 6.6706 3.9589 13.8973 27.9841   NA  15
#> phi-sp2      19.0082 6.8560 6.3426 20.4508 28.4248   NA  36
#> phi-sp3      13.0627 7.1832 3.5532 11.8307 26.5863   NA  21
#> phi-sp4      15.7934 7.0737 4.1299 15.1926 28.6472   NA  20
#> phi-sp5      19.3351 7.0978 6.4903 19.7914 29.5367   NA  24

# Predict at new locations ------------------------------------------------
out.pred <- predict(out, X.0, coords.0, verbose = FALSE)