Function for prediction at new locations for spatially varying coefficient multi-species occupancy models

The function predict collects posterior predictive samples for a set of new locations given an object of class `svcMsPGOcc`. Prediction is possible for both the latent occupancy state as well as detection.

Usage

# S3 method for svcMsPGOcc
predict(object, X.0, coords.0, n.omp.threads = 1, verbose = TRUE, 
        n.report = 100, ignore.RE = FALSE, type = 'occupancy', ...)

Arguments

object

an object of class svcMsPGOcc

X.0

the design matrix of covariates at the prediction locations. This should include a column of 1s for the intercept if an intercept is included in the model. If random effects are included in the occupancy (or detection if type = 'detection') portion of the model, the levels of the random effects at the new locations should be included as a column in the design matrix. The ordering of the levels should match the ordering used to fit the data in svcMsPGOcc. Columns should correspond to the order of how covariates were specified in the corresponding formula argument of svcMsPGOcc. Column names of the random effects must match the name of the random effects, if specified in the corresponding formula argument of svcMsPGOcc.

coords.0

the spatial coordinates corresponding to X.0. Note that spOccupancy assumes coordinates are specified in a projected coordinate system.

n.omp.threads

a positive integer indicating the number of threads to use for SMP parallel processing. The package must be compiled for OpenMP support. For most Intel-based machines, we recommend setting n.omp.threads up to the number of hyperthreaded cores. Note, n.omp.threads > 1 might not work on some systems.

verbose

if TRUE, model specification and progress of the sampler is printed to the screen. Otherwise, nothing is printed to the screen.

n.report

the interval to report sampling progress.

ignore.RE

a logical value indicating whether to include unstructured random effects for prediction. If TRUE, unstructured random effects will be ignored and prediction will only use the fixed effects and the spatial random effects. If FALSE, random effects will be included in the prediction for both observed and unobserved levels of the unstructured random effects.

type

a quoted keyword indicating what type of prediction to produce. Valid keywords are 'occupancy' to predict latent occupancy probability and latent occupancy values (this is the default), or 'detection' to predict detection probability given new values of detection covariates.

...

currently no additional arguments

Note

When ignore.RE = FALSE, both sampled levels and non-sampled levels of random effects are supported for prediction. For sampled levels, the posterior distribution for the random intercept corresponding to that level of the random effect will be used in the prediction. For non-sampled levels, random values are drawn from a normal distribution using the posterior samples of the random effect variance, which results in fully propagated uncertainty in predictions with models that incorporate random effects.

Author

Jeffrey W. Doser doserjef@msu.edu,
Andrew O. Finley finleya@msu.edu

Value

An list object of class predict.svcMsPGOcc. When type = 'occupancy', the list consists of:

psi.0.samples

a three-dimensional array of posterior predictive samples for the latent occurrence probability values.

z.0.samples

a three-dimensional array of posterior predictive samples for the latent occurrence values.

w.0.samples

a four-dimensional array of posterior predictive samples for the spatially-varying coefficients, with dimensions corresponding to MCMC sample, spatial factor, site, and spatially varying coefficient.

run.time

execution time reported using proc.time().

When type = 'detection', the list consists of:

p.0.samples

a three-dimensional array of posterior predictive samples for the detection probability values.

run.time

execution time reported using proc.time().

The return object will include additional objects used for standard extractor functions.

Examples

set.seed(400)

# Simulate Data -----------------------------------------------------------
J.x <- 10
J.y <- 10
J <- J.x * J.y
n.rep <- sample(5, size = J, replace = TRUE)
N <- 6
# Community-level covariate effects
# Occurrence
beta.mean <- c(0.2, -0.2, 0.3, -0.1, 0.4)
p.occ <- length(beta.mean)
tau.sq.beta <- c(0.6, 1.5, 0.4, 0.5, 0.3)
# Detection
alpha.mean <- c(0, 1.2, -0.5)
tau.sq.alpha <- c(1, 0.5, 1.3)
p.det <- length(alpha.mean)
# No random effects
psi.RE <- list()
p.RE <- list()
# Draw species-level effects from community means.
beta <- matrix(NA, nrow = N, ncol = p.occ)
alpha <- matrix(NA, nrow = N, ncol = p.det)
for (i in 1:p.occ) {
  beta[, i] <- rnorm(N, beta.mean[i], sqrt(tau.sq.beta[i]))
}
for (i in 1:p.det) {
  alpha[, i] <- rnorm(N, alpha.mean[i], sqrt(tau.sq.alpha[i]))
}
# Number of spatial factors for each SVC
n.factors <- 2
# The intercept and first two covariates have spatially-varying effects
svc.cols <- c(1, 2, 3)
p.svc <- length(svc.cols)
q.p.svc <- n.factors * p.svc
# Spatial decay parameters
phi <- runif(q.p.svc, 3 / 0.9, 3 / 0.1)
# A length N vector indicating the proportion of simulated locations
# that are within the range for a given species.
range.probs <- runif(N, 1, 1)
factor.model <- TRUE
cov.model <- 'spherical'
sp <- TRUE

dat <- simMsOcc(J.x = J.x, J.y = J.y, n.rep = n.rep, N = N, beta = beta, alpha = alpha,
    psi.RE = psi.RE, p.RE = p.RE, phi = phi, sp = sp, svc.cols = svc.cols,
    cov.model = cov.model, n.factors = n.factors,
    factor.model = factor.model, range.probs = range.probs)

# Split into fitting and prediction data set
pred.indx <- sample(1:J, round(J * .25), replace = FALSE)
y <- dat$y[, -pred.indx, ]
# Occupancy covariates
X <- dat$X[-pred.indx, ]
# Coordinates
coords <- as.matrix(dat$coords[-pred.indx, ])
# Detection covariates
X.p <- dat$X.p[-pred.indx, , ]
# Prediction values
X.0 <- dat$X[pred.indx, ]
coords.0 <- as.matrix(dat$coords[pred.indx, ])

# Prep data for spOccupancy -----------------------------------------------
# Occurrence covariates
occ.covs <- cbind(X)
colnames(occ.covs) <- c('int', 'occ.cov.1', 'occ.cov.2', 'occ.cov.3',
      'occ.cov.4')
# Detection covariates
det.covs <- list(det.cov.1 = X.p[, , 2],
     det.cov.2 = X.p[, , 3])
# Data list
data.list <- list(y = y, coords = coords, occ.covs = occ.covs,
                  det.covs = det.covs)
# Priors
prior.list <- list(beta.comm.normal = list(mean = 0, var = 2.72),
       alpha.comm.normal = list(mean = 0, var = 2.72),
       tau.sq.beta.ig = list(a = 0.1, b = 0.1),
       tau.sq.alpha.ig = list(a = 0.1, b = 0.1),
                   phi.unif = list(a = 3 / 1, b = 3 / .1))
inits.list <- list(alpha.comm = 0,
       beta.comm = 0,
       beta = 0,
       alpha = 0,
       tau.sq.beta = 1,
       tau.sq.alpha = 1,
       z = apply(y, c(1, 2), max, na.rm = TRUE))
# Tuning
tuning.list <- list(phi = 1)

# Number of batches
n.batch <- 2
# Batch length
batch.length <- 25
n.burn <- 0
n.thin <- 1
n.samples <- n.batch * batch.length

out <- svcMsPGOcc(occ.formula = ~ occ.cov.1 + occ.cov.2 + occ.cov.3 +
                                  occ.cov.4,
                  det.formula = ~ det.cov.1 + det.cov.2,
                  data = data.list,
                  inits = inits.list,
                  n.batch = n.batch,
      n.factors = n.factors,
                  batch.length = batch.length,
      std.by.sp = TRUE,
                  accept.rate = 0.43,
                  priors = prior.list,
      svc.cols = svc.cols,
                  cov.model = "spherical",
                  tuning = tuning.list,
                  n.omp.threads = 1,
                  verbose = TRUE,
                  NNGP = TRUE,
                  n.neighbors = 5,
                  search.type = 'cb',
                  n.report = 10,
                  n.burn = n.burn,
      n.thin = n.thin,
      n.chains = 1)
#> ----------------------------------------
#> 	Preparing to run the model
#> ----------------------------------------
#> phi is not specified in initial values.
#> Setting initial value to random values from the prior distribution
#> lambda is not specified in initial values.
#> Setting initial values of the lower triangle to random values from a standard normal
#> lambda is not specified in initial values.
#> Setting initial values of the lower triangle to random values from a standard normal
#> w is not specified in initial values.
#> Setting initial value to 0
#> ----------------------------------------
#> 	Building the neighbor list
#> ----------------------------------------
#> ----------------------------------------
#> Building the neighbors of neighbors list
#> ----------------------------------------
#> ----------------------------------------
#> 	Model description
#> ----------------------------------------
#> Spatial Factor NNGP Multispecies Occupancy Model with Polya-Gamma latent
#> variable fit with 75 sites and 6 species.
#> 
#> Samples per chain: 50 (2 batches of length 25)
#> Burn-in: 0 
#> Thinning Rate: 1 
#> Number of Chains: 1 
#> Total Posterior Samples: 50 
#> 
#> Number of spatially-varying coefficients: 3 
#> Using the spherical spatial correlation model.
#> 
#> Using 2 latent spatial factors.
#> Using 5 nearest neighbors.
#> 
#> Source compiled with OpenMP support and model fit using 1 thread(s).
#> 
#> Adaptive Metropolis with target acceptance rate: 43.0
#> ----------------------------------------
#> 	Chain 1
#> ----------------------------------------
#> Sampling ... 
#> Batch: 2 of 2, 100.00%

summary(out)
#> 
#> Call:
#> svcMsPGOcc(occ.formula = ~occ.cov.1 + occ.cov.2 + occ.cov.3 + 
#>     occ.cov.4, det.formula = ~det.cov.1 + det.cov.2, data = data.list, 
#>     inits = inits.list, priors = prior.list, tuning = tuning.list, 
#>     svc.cols = svc.cols, cov.model = "spherical", NNGP = TRUE, 
#>     n.neighbors = 5, search.type = "cb", std.by.sp = TRUE, n.factors = n.factors, 
#>     n.batch = n.batch, batch.length = batch.length, accept.rate = 0.43, 
#>     n.omp.threads = 1, verbose = TRUE, n.report = 10, n.burn = n.burn, 
#>     n.thin = n.thin, n.chains = 1)
#> 
#> Samples per Chain: 50
#> Burn-in: 0
#> Thinning Rate: 1
#> Number of Chains: 1
#> Total Posterior Samples: 50
#> Run Time (min): 0.0081
#> 
#> ----------------------------------------
#> 	Community Level
#> ----------------------------------------
#> Occurrence Means (logit scale): 
#>               Mean     SD    2.5%    50%  97.5% Rhat ESS
#> (Intercept) 0.1144 0.4904 -0.8193 0.1938 0.8416   NA  10
#> occ.cov.1   0.1528 0.4919 -0.6581 0.0951 1.1168   NA  50
#> occ.cov.2   0.5542 0.4759 -0.1257 0.5031 1.2628   NA   9
#> occ.cov.3   0.0773 0.2669 -0.3549 0.0656 0.5963   NA  50
#> occ.cov.4   0.5731 0.3984 -0.2794 0.5899 1.2895   NA  18
#> 
#> Occurrence Variances (logit scale): 
#>               Mean     SD   2.5%    50%  97.5% Rhat ESS
#> (Intercept) 1.2137 1.2112 0.2175 0.7563 4.3797   NA  12
#> occ.cov.1   1.9187 1.5023 0.1924 1.6037 5.2332   NA  18
#> occ.cov.2   0.3834 0.3662 0.0535 0.2924 1.4950   NA  25
#> occ.cov.3   0.4702 0.5472 0.0693 0.3213 1.6088   NA  50
#> occ.cov.4   0.3902 0.3350 0.0919 0.3012 1.4657   NA  35
#> 
#> Detection Means (logit scale): 
#>                Mean     SD    2.5%     50%  97.5% Rhat ESS
#> (Intercept) -0.1186 0.2890 -0.7290 -0.1235 0.3319   NA  50
#> det.cov.1    1.2844 0.4581  0.4751  1.2783 2.2285   NA  50
#> det.cov.2   -0.3115 0.4466 -1.0890 -0.3079 0.6505   NA 105
#> 
#> Detection Variances (logit scale): 
#>               Mean     SD   2.5%    50%  97.5% Rhat ESS
#> (Intercept) 0.7481 0.6742 0.1697 0.5334 2.9542   NA  50
#> det.cov.1   0.7040 0.5456 0.0830 0.6282 2.0466   NA  23
#> det.cov.2   2.2632 2.1346 0.1349 1.5215 8.7636   NA  50
#> 
#> ----------------------------------------
#> 	Species Level
#> ----------------------------------------
#> Occurrence (logit scale): 
#>                    Mean     SD    2.5%     50%   97.5% Rhat ESS
#> (Intercept)-sp1  0.3250 0.3720 -0.3171  0.3253  1.1011   NA  19
#> (Intercept)-sp2  0.2434 0.4454 -0.4878  0.1996  1.1363   NA  15
#> (Intercept)-sp3  1.4276 0.8155  0.4118  1.2886  3.2438   NA   3
#> (Intercept)-sp4  0.0076 0.5066 -0.6044 -0.0828  1.0716   NA   6
#> (Intercept)-sp5  0.2147 0.5427 -0.6454  0.2075  0.9976   NA   6
#> (Intercept)-sp6 -1.1061 0.5100 -2.1891 -1.1147 -0.2510   NA   5
#> occ.cov.1-sp1    0.3709 0.4718 -0.5855  0.4243  1.1807   NA  18
#> occ.cov.1-sp2    2.2318 0.7893  0.4333  2.3542  3.3875   NA   5
#> occ.cov.1-sp3    0.6679 0.5434 -0.1913  0.6213  1.9709   NA   7
#> occ.cov.1-sp4   -0.9249 0.5121 -1.8189 -0.8693 -0.2072   NA  11
#> occ.cov.1-sp5   -0.5175 0.5209 -1.3868 -0.4688  0.3743   NA   5
#> occ.cov.1-sp6   -0.5846 0.5410 -1.4711 -0.5532  0.3441   NA   9
#> occ.cov.2-sp1    0.4385 0.4327 -0.5150  0.3777  1.1035   NA  11
#> occ.cov.2-sp2    0.8307 0.4658  0.1969  0.7145  1.8925   NA   7
#> occ.cov.2-sp3    0.4332 0.5853 -0.3927  0.3684  2.0287   NA   5
#> occ.cov.2-sp4    0.4616 0.5766 -0.5239  0.3961  1.4415   NA   5
#> occ.cov.2-sp5    1.1355 0.4700  0.2772  1.1171  1.9099   NA   9
#> occ.cov.2-sp6    0.3084 0.4716 -0.4433  0.2735  1.2479   NA   8
#> occ.cov.3-sp1    0.2744 0.3948 -0.4248  0.2345  0.9369   NA   6
#> occ.cov.3-sp2   -0.1037 0.3559 -0.7343 -0.1353  0.4760   NA  11
#> occ.cov.3-sp3   -0.4159 0.3550 -1.1208 -0.4062  0.1249   NA  12
#> occ.cov.3-sp4    0.6729 0.4138 -0.0859  0.5846  1.3742   NA   9
#> occ.cov.3-sp5    0.0336 0.4167 -0.6957 -0.0180  0.9220   NA   9
#> occ.cov.3-sp6    0.0699 0.3919 -0.5499  0.0224  0.8498   NA   9
#> occ.cov.4-sp1    1.0989 0.3866  0.4563  1.0764  1.7141   NA  23
#> occ.cov.4-sp2    0.7053 0.4145  0.0000  0.6802  1.4505   NA  13
#> occ.cov.4-sp3    0.0503 0.4802 -1.0673  0.1141  0.8687   NA  36
#> occ.cov.4-sp4    0.7495 0.4399 -0.1047  0.6821  1.5778   NA  11
#> occ.cov.4-sp5    0.5502 0.3082  0.0170  0.5364  1.1194   NA  21
#> occ.cov.4-sp6    0.6334 0.4578 -0.0206  0.6357  1.7078   NA  14
#> 
#> Detection (logit scale): 
#>                    Mean     SD    2.5%     50%   97.5% Rhat ESS
#> (Intercept)-sp1  0.8174 0.2507  0.4624  0.7841  1.3856   NA  26
#> (Intercept)-sp2  0.1528 0.2702 -0.2657  0.1003  0.7316   NA  28
#> (Intercept)-sp3  0.3354 0.1668  0.0508  0.3387  0.6990   NA  28
#> (Intercept)-sp4 -0.9590 0.2678 -1.4780 -0.9314 -0.4387   NA  15
#> (Intercept)-sp5 -0.7488 0.2574 -1.2021 -0.7281 -0.3116   NA   8
#> (Intercept)-sp6 -0.4568 0.3060 -1.0343 -0.4212  0.1291   NA  23
#> det.cov.1-sp1    1.0731 0.2253  0.6548  1.0806  1.5254   NA  56
#> det.cov.1-sp2    2.2921 0.5199  1.0962  2.3359  3.0576   NA  10
#> det.cov.1-sp3    0.4349 0.1568  0.1582  0.4278  0.6770   NA  50
#> det.cov.1-sp4    1.6887 0.3293  1.0418  1.7016  2.2648   NA  26
#> det.cov.1-sp5    1.2684 0.2441  0.9558  1.2756  1.6561   NA  25
#> det.cov.1-sp6    1.3766 0.3679  0.6494  1.4053  1.9358   NA  18
#> det.cov.2-sp1   -0.6310 0.2736 -1.2233 -0.5617 -0.2219   NA  36
#> det.cov.2-sp2   -1.9175 0.5066 -2.7712 -1.9048 -1.0406   NA  14
#> det.cov.2-sp3   -0.9691 0.2167 -1.3326 -0.9837 -0.4591   NA  26
#> det.cov.2-sp4   -0.2251 0.2060 -0.6949 -0.2583  0.2252   NA  28
#> det.cov.2-sp5    0.8162 0.3087  0.1550  0.8520  1.4132   NA   9
#> det.cov.2-sp6    0.9893 0.3799  0.0858  0.9770  1.5348   NA  16
#> 
#> ----------------------------------------
#> 	Spatial Covariance
#> ----------------------------------------
#>                      Mean     SD    2.5%     50%   97.5% Rhat ESS
#> phi-1-(Intercept) 17.4972 7.4417  5.3614 16.2150 28.8437   NA   4
#> phi-2-(Intercept) 16.5271 5.3242  7.2684 16.9005 25.5660   NA  15
#> phi-1-occ.cov.1   18.5748 5.4079  9.7565 17.4329 27.7051   NA   9
#> phi-2-occ.cov.1   12.4520 5.9112  6.6358  9.5261 25.4593   NA   8
#> phi-1-occ.cov.2   19.8195 5.4046 10.8523 20.9929 26.7767   NA   6
#> phi-2-occ.cov.2   17.1302 6.8536  6.7466 14.9697 27.6486   NA   6
# Predict at new locations ------------------------------------------------
out.pred <- predict(out, X.0, coords.0, verbose = FALSE)

# Get SVC samples for each species at prediction locations
svc.samples <- getSVCSamples(out, out.pred)