Function for prediction at new locations for spatial factor multi-species occupancy models

The function predict collects posterior predictive samples for a set of new locations given an object of class `sfMsPGOcc`. Prediction is possible for both the latent occupancy state as well as detection.

# S3 method for sfMsPGOcc
predict(object, X.0, coords.0, n.omp.threads = 1, verbose = TRUE, 
        n.report = 100, ignore.RE = FALSE, type = 'occupancy', ...)

Arguments

object

an object of class sfMsPGOcc

X.0

the design matrix of covariates at the prediction locations. This should include a column of 1s for the intercept if an intercept is included in the model. If random effects are included in the occupancy (or detection if type = 'detection') portion of the model, the levels of the random effects at the new locations should be included as a column in the design matrix. The ordering of the levels should match the ordering used to fit the data in sfMsPGOcc. Columns should correspond to the order of how covariates were specified in the corresponding formula argument of sfMsPGOcc. Column names of the random effects must match the name of the random effects, if specified in the corresponding formula argument of sfMsPGOcc.

coords.0

the spatial coordinates corresponding to X.0. Note that spOccupancy assumes coordinates are specified in a projected coordinate system.

n.omp.threads

a positive integer indicating the number of threads to use for SMP parallel processing. The package must be compiled for OpenMP support. For most Intel-based machines, we recommend setting n.omp.threads up to the number of hyperthreaded cores. Note, n.omp.threads > 1 might not work on some systems.

verbose

if TRUE, model specification and progress of the sampler is printed to the screen. Otherwise, nothing is printed to the screen.

n.report

the interval to report sampling progress.

ignore.RE

a logical value indicating whether to include unstructured random effects for prediction. If TRUE, unstructured random effects will be ignored and prediction will only use the fixed effects and the spatial random effects. If FALSE, random effects will be included in the prediction for both observed and unobserved levels of the unstructured random effects.

type

a quoted keyword indicating what type of prediction to produce. Valid keywords are 'occupancy' to predict latent occupancy probability and latent occupancy values (this is the default), or 'detection' to predict detection probability given new values of detection covariates.

...

currently no additional arguments

Note

When ignore.RE = FALSE, both sampled levels and non-sampled levels of random effects are supported for prediction. For sampled levels, the posterior distribution for the random intercept corresponding to that level of the random effect will be used in the prediction. For non-sampled levels, random values are drawn from a normal distribution using the posterior samples of the random effect variance, which results in fully propagated uncertainty in predictions with models that incorporate random effects.

Author

Jeffrey W. Doser doserjef@msu.edu,
Andrew O. Finley finleya@msu.edu

Value

An list object of class predict.sfMsPGOcc. When type = 'occupancy', the list consists of:

psi.0.samples

a three-dimensional array of posterior predictive samples for the latent occurrence probability values.

z.0.samples

a three-dimensional array of posterior predictive samples for the latent occurrence values.

w.0.samples

a three-dimensional array of posterior predictive samples for the latent spatial factors.

run.time

execution time reported using proc.time().

When type = 'detection', the list consists of:

p.0.samples

a three-dimensional array of posterior predictive samples for the detection probability values.

run.time

execution time reported using proc.time().

The return object will include additional objects used for standard extractor functions.

Examples

set.seed(400)

# Simulate Data -----------------------------------------------------------
J.x <- 7
J.y <- 7
J <- J.x * J.y
n.rep <- sample(2:4, size = J, replace = TRUE)
N <- 5
# Community-level covariate effects
# Occurrence
beta.mean <- c(0.2, -0.15)
p.occ <- length(beta.mean)
tau.sq.beta <- c(0.6, 0.3)
# Detection
alpha.mean <- c(0.5, 0.2, -.2)
tau.sq.alpha <- c(0.2, 0.3, 0.8)
p.det <- length(alpha.mean)
# Draw species-level effects from community means.
beta <- matrix(NA, nrow = N, ncol = p.occ)
alpha <- matrix(NA, nrow = N, ncol = p.det)
for (i in 1:p.occ) {
  beta[, i] <- rnorm(N, beta.mean[i], sqrt(tau.sq.beta[i]))
}
for (i in 1:p.det) {
  alpha[, i] <- rnorm(N, alpha.mean[i], sqrt(tau.sq.alpha[i]))
}
n.factors <- 3
phi <- runif(n.factors, 3/1, 3/.4)
sp <- TRUE

dat <- simMsOcc(J.x = J.x, J.y = J.y, n.rep = n.rep, N = N, beta = beta, alpha = alpha,
                phi = phi, sigma.sq = sigma.sq, sp = TRUE, cov.model = 'exponential', 
                factor.model = TRUE, n.factors = n.factors)
#> sigma.sq is specified but will be set to 1 for spatial latent factor model

# Number of batches
n.batch <- 10
# Batch length
batch.length <- 25
n.samples <- n.batch * batch.length

# Split into fitting and prediction data set
pred.indx <- sample(1:J, round(J * .25), replace = FALSE)
y <- dat$y[, -pred.indx, ]
# Occupancy covariates
X <- dat$X[-pred.indx, ]
# Coordinates
coords <- as.matrix(dat$coords[-pred.indx, ])
# Detection covariates
X.p <- dat$X.p[-pred.indx, , ]
# Prediction values
X.0 <- dat$X[pred.indx, ]
coords.0 <- as.matrix(dat$coords[pred.indx, ])
psi.0 <- dat$psi[, pred.indx]

# Package all data into a list
occ.covs <- X[, 2, drop = FALSE]
colnames(occ.covs) <- c('occ.cov')
det.covs <- list(det.cov.1 = X.p[, , 2], 
                 det.cov.2 = X.p[, , 3])
data.list <- list(y = y, 
                  occ.covs = occ.covs,
                  det.covs = det.covs, 
                  coords = coords)

# Priors 
prior.list <- list(beta.comm.normal = list(mean = 0, var = 2.72), 
                   alpha.comm.normal = list(mean = 0, var = 2.72), 
                   tau.sq.beta.ig = list(a = 0.1, b = 0.1), 
                   tau.sq.alpha.ig = list(a = 0.1, b = 0.1),
                   phi.unif = list(a = 3/1, b = 3/.1), 
                   sigma.sq.ig = list(a = 2, b = 2)) 
# Starting values
lambda.inits <- matrix(0, N, n.factors)
diag(lambda.inits) <- 1
lambda.inits[lower.tri(lambda.inits)] <- rnorm(sum(lower.tri(lambda.inits)))
inits.list <- list(alpha.comm = 0, 
                   beta.comm = 0, 
                   beta = 0, 
                   alpha = 0,
                   tau.sq.beta = 1, 
                   tau.sq.alpha = 1, 
                   phi = 3 / .5, 
                   sigma.sq = 2,
                   lambda = lambda.inits,
                   z = apply(y, c(1, 2), max, na.rm = TRUE))
# Tuning
tuning.list <- list(phi = 1) 

out <- sfMsPGOcc(occ.formula = ~ occ.cov, 
                 det.formula = ~ det.cov.1 + det.cov.2, 
                 data = data.list,
                 inits = inits.list, 
                 n.batch = n.batch, 
                 batch.length = batch.length, 
                 accept.rate = 0.43, 
                 n.factors = 3,
                 priors = prior.list, 
                 cov.model = "exponential", 
                 tuning = tuning.list, 
                 n.omp.threads = 1, 
                 verbose = TRUE, 
                 NNGP = TRUE, 
                 n.neighbors = 5, 
                 search.type = 'cb', 
                 n.report = 10, 
                 n.burn = 100, 
                 n.thin = 1)
#> ----------------------------------------
#> 	Preparing to run the model
#> ----------------------------------------
#> ----------------------------------------
#> 	Building the neighbor list
#> ----------------------------------------
#> ----------------------------------------
#> Building the neighbors of neighbors list
#> ----------------------------------------
#> ----------------------------------------
#> 	Model description
#> ----------------------------------------
#> Spatial Factor NNGP Multispecies Occupancy Model with Polya-Gamma latent
#> variable fit with 37 sites and 5 species.
#> 
#> Samples per chain: 250 (10 batches of length 25)
#> Burn-in: 100 
#> Thinning Rate: 1 
#> Number of Chains: 1 
#> Total Posterior Samples: 150 
#> 
#> Using the exponential spatial correlation model.
#> 
#> Using 3 latent spatial factors.
#> Using 5 nearest neighbors.
#> 
#> Source compiled with OpenMP support and model fit using 1 thread(s).
#> 
#> Adaptive Metropolis with target acceptance rate: 43.0
#> ----------------------------------------
#> 	Chain 1
#> ----------------------------------------
#> Sampling ... 
#> Batch: 10 of 10, 100.00%

summary(out, level = 'both')
#> 
#> Call:
#> sfMsPGOcc(occ.formula = ~occ.cov, det.formula = ~det.cov.1 + 
#>     det.cov.2, data = data.list, inits = inits.list, priors = prior.list, 
#>     tuning = tuning.list, cov.model = "exponential", NNGP = TRUE, 
#>     n.neighbors = 5, search.type = "cb", n.factors = 3, n.batch = n.batch, 
#>     batch.length = batch.length, accept.rate = 0.43, n.omp.threads = 1, 
#>     verbose = TRUE, n.report = 10, n.burn = 100, n.thin = 1)
#> 
#> Samples per Chain: 250
#> Burn-in: 100
#> Thinning Rate: 1
#> Number of Chains: 1
#> Total Posterior Samples: 150
#> Run Time (min): 0.0061
#> 
#> ----------------------------------------
#> 	Community Level
#> ----------------------------------------
#> Occurrence Means (logit scale): 
#>                Mean     SD    2.5%     50%  97.5% Rhat ESS
#> (Intercept)  0.1412 0.5560 -1.0437  0.1602 1.0910   NA 150
#> occ.cov     -0.3226 0.5592 -1.5968 -0.3414 0.8973   NA  58
#> 
#> Occurrence Variances (logit scale): 
#>               Mean     SD   2.5%    50%  97.5% Rhat ESS
#> (Intercept) 1.4863 1.8645 0.0948 0.7959 7.0631   NA  47
#> occ.cov     1.7294 3.0577 0.0774 0.7874 7.8695   NA  87
#> 
#> Detection Means (logit scale): 
#>                Mean     SD    2.5%     50%   97.5% Rhat ESS
#> (Intercept)  0.7752 0.3462  0.0505  0.8282  1.3649   NA  65
#> det.cov.1   -0.1104 0.3165 -0.6833 -0.1001  0.4499   NA  74
#> det.cov.2   -0.8464 0.3786 -1.4786 -0.8557 -0.1489   NA 111
#> 
#> Detection Variances (logit scale): 
#>               Mean     SD   2.5%    50%  97.5% Rhat ESS
#> (Intercept) 0.3359 0.4262 0.0387 0.1880 1.5742   NA  34
#> det.cov.1   0.5189 0.7485 0.0608 0.2596 3.1192   NA  66
#> det.cov.2   0.7576 1.0870 0.0833 0.4867 3.0461   NA  61
#> 
#> ----------------------------------------
#> 	Species Level
#> ----------------------------------------
#> Occurrence (logit scale): 
#>                    Mean     SD    2.5%     50%   97.5% Rhat ESS
#> (Intercept)-sp1  0.4236 0.3946 -0.2295  0.4109  1.1871   NA  95
#> (Intercept)-sp2  0.4555 0.4510 -0.4973  0.4715  1.3453   NA  51
#> (Intercept)-sp3  0.3178 0.5086 -0.6081  0.3384  1.3754   NA  38
#> (Intercept)-sp4  0.4507 0.4811 -0.4961  0.4665  1.4518   NA  45
#> (Intercept)-sp5 -1.0409 0.8443 -3.0321 -0.9679  0.2676   NA  17
#> occ.cov-sp1     -0.3835 0.3505 -1.1504 -0.3567  0.2688   NA  38
#> occ.cov-sp2     -0.7864 0.4677 -1.7710 -0.7128 -0.0261   NA  48
#> occ.cov-sp3     -0.5599 0.4721 -1.4477 -0.5673  0.3831   NA  72
#> occ.cov-sp4      1.0363 0.8359 -0.2308  0.9527  3.0909   NA  12
#> occ.cov-sp5     -0.8581 0.5167 -2.0289 -0.7614 -0.0970   NA  45
#> 
#> Detection (logit scale): 
#>                    Mean     SD    2.5%     50%   97.5% Rhat ESS
#> (Intercept)-sp1  1.1679 0.2902  0.7047  1.1578  1.8023   NA  88
#> (Intercept)-sp2  0.9227 0.2869  0.3750  0.9155  1.4633   NA  61
#> (Intercept)-sp3  0.6117 0.3218 -0.0478  0.6462  1.2256   NA  51
#> (Intercept)-sp4  0.5145 0.2842 -0.0452  0.5200  1.0053   NA  49
#> (Intercept)-sp5  0.7225 0.4388 -0.4509  0.7171  1.3981   NA  38
#> det.cov.1-sp1    0.0522 0.2613 -0.4454  0.0590  0.6090   NA  95
#> det.cov.1-sp2   -0.4470 0.2667 -1.0010 -0.4298 -0.0365   NA  82
#> det.cov.1-sp3    0.5411 0.3542 -0.1334  0.5162  1.2617   NA  48
#> det.cov.1-sp4   -0.4360 0.3183 -1.1151 -0.4240  0.1411   NA 141
#> det.cov.1-sp5   -0.1228 0.3359 -0.8469 -0.1132  0.4630   NA  78
#> det.cov.2-sp1   -0.4280 0.2705 -0.9079 -0.4326  0.0712   NA  70
#> det.cov.2-sp2   -0.6808 0.2999 -1.3164 -0.6656 -0.1337   NA 120
#> det.cov.2-sp3   -1.8891 0.4639 -2.7881 -1.8394 -1.1535   NA  25
#> det.cov.2-sp4   -0.7493 0.2637 -1.2686 -0.7108 -0.2809   NA 130
#> det.cov.2-sp5   -0.8094 0.3814 -1.6185 -0.7627 -0.1914   NA 109
#> 
#> ----------------------------------------
#> 	Spatial Covariance
#> ----------------------------------------
#>          Mean     SD   2.5%     50%   97.5% Rhat ESS
#> phi-1 18.0211 7.4472 5.9352 18.3516 29.1761   NA  14
#> phi-2 15.8601 7.1367 5.1977 16.9432 27.7161   NA  13
#> phi-3 15.0597 7.0939 4.5686 15.6574 28.3604   NA  10

# Predict at new locations ------------------------------------------------
out.pred <- predict(out, X.0, coords.0, verbose = FALSE)