Function for prediction at new locations for single species integrated spatial occupancy models

The function predict collects posterior predictive samples for a set of new locations given an object of class `spIntPGOcc`.

# S3 method for spIntPGOcc
predict(object, X.0, coords.0, n.omp.threads = 1, verbose = TRUE, 
                             n.report = 100, ...)

Arguments

object
an object of class spIntPGOcc.

X.0
the design matrix for prediction locations. This should include a column of 1s for the intercept. Covariates should have the same column names as those used when fitting the model with spIntPGOcc.

coords.0
the spatial coordinates corresponding to X.0.

n.omp.threads
a positive integer indicating the number of threads to use for SMP parallel processing. The package must be compiled for OpenMP support. For most Intel-based machines, we recommend setting n.omp.threads up to the number of hyperthreaded cores. Note, n.omp.threads > 1 might not work on some systems.

verbose
if TRUE, model specification and progress of the sampler is printed to the screen. Otherwise, nothing is printed to the screen.

n.report
the interval to report sampling progress.

...
currently no additional arguments

Author

Jeffrey W. Doser doserjef@msu.edu,
Andrew O. Finley finleya@msu.edu

Value

An object of class predict.spIntPGOcc that is a list comprised of:

psi.0.samples

a coda object of posterior predictive samples for the latent occurrence probability values.

z.0.samples

a coda object of posterior predictive samples for the latent occurrence values.

The return object will include additional objects used for standard extractor functions.
References

Hooten, M. B., and Hefley, T. J. (2019). Bringing Bayesian models to life. CRC Press.

Examples

set.seed(400)

# Simulate Data -----------------------------------------------------------
# Number of locations in each direction. This is the total region of interest
# where some sites may or may not have a data source. 
J.x <- 8
J.y <- 8
J.all <- J.x * J.y
# Number of data sources.
n.data <- 4
# Sites for each data source. 
J.obs <- sample(ceiling(0.2 * J.all):ceiling(0.5 * J.all), n.data, replace = TRUE)
# Replicates for each data source.
n.rep <- list()
for (i in 1:n.data) {
  n.rep[[i]] <- sample(1:4, size = J.obs[i], replace = TRUE)
}
# Occupancy covariates
beta <- c(0.5, 0.5)
p.occ <- length(beta)
# Detection covariates
alpha <- list()
alpha[[1]] <- runif(2, 0, 1)
alpha[[2]] <- runif(3, 0, 1)
alpha[[3]] <- runif(2, -1, 1)
alpha[[4]] <- runif(4, -1, 1)
p.det.long <- sapply(alpha, length)
p.det <- sum(p.det.long)
sigma.sq <- 2
phi <- 3 / .5
sp <- TRUE

# Simulate occupancy data. 
dat <- simIntOcc(n.data = n.data, J.x = J.x, J.y = J.y, J.obs = J.obs, 
     n.rep = n.rep, beta = beta, alpha = alpha, sp = sp, 
     phi = phi, sigma.sq = sigma.sq, cov.model = 'spherical')

y <- dat$y
X <- dat$X.obs
X.p <- dat$X.p
sites <- dat$sites
X.0 <- dat$X.pred
psi.0 <- dat$psi.pred
coords <- as.matrix(dat$coords.obs)
coords.0 <- as.matrix(dat$coords.pred)

# Package all data into a list
occ.covs <- X[, 2, drop = FALSE]
colnames(occ.covs) <- c('occ.cov')
det.covs <- list()
# Add covariates one by one
det.covs[[1]] <- list(det.cov.1.1 = X.p[[1]][, , 2])
det.covs[[2]] <- list(det.cov.2.1 = X.p[[2]][, , 2], 
          det.cov.2.2 = X.p[[2]][, , 3])
det.covs[[3]] <- list(det.cov.3.1 = X.p[[3]][, , 2])
det.covs[[4]] <- list(det.cov.4.1 = X.p[[4]][, , 2], 
          det.cov.4.2 = X.p[[4]][, , 3], 
          det.cov.4.3 = X.p[[4]][, , 4])
data.list <- list(y = y, 
      occ.covs = occ.covs,
      det.covs = det.covs, 
      sites = sites, 
      coords = coords)

J <- length(dat$z.obs)

# Initial values
inits.list <- list(alpha = list(0, 0, 0, 0), 
          beta = 0, 
          phi = 3 / .5, 
          sigma.sq = 2, 
          w = rep(0, J), 
          z = rep(1, J))
# Priors
prior.list <- list(beta.normal = list(mean = 0, var = 2.72), 
       alpha.normal = list(mean = list(0, 0, 0, 0), 
               var = list(2.72, 2.72, 2.72, 2.72)),
       phi.unif = c(3/1, 3/.1), 
       sigma.sq.ig = c(2, 2))
# Tuning
tuning.list <- list(phi = 1) 

# Number of batches
n.batch <- 40
# Batch length
batch.length <- 25

out <- spIntPGOcc(occ.formula = ~ occ.cov, 
      det.formula = list(f.1 = ~ det.cov.1.1, 
             f.2 = ~ det.cov.2.1 + det.cov.2.2, 
             f.3 = ~ det.cov.3.1, 
             f.4 = ~ det.cov.4.1 + det.cov.4.2 + det.cov.4.3), 
      data = data.list,  
      inits = inits.list, 
      n.batch = n.batch, 
      batch.length = batch.length, 
      accept.rate = 0.43, 
      priors = prior.list, 
      cov.model = "spherical", 
      tuning = tuning.list, 
      n.omp.threads = 1, 
      verbose = TRUE, 
      NNGP = TRUE, 
      n.neighbors = 5, 
      search.type = 'cb', 
      n.report = 10, 
      n.burn = 500, 
      n.thin = 1)
#> ----------------------------------------
#> 	Preparing the data
#> ----------------------------------------
#> ----------------------------------------
#> 	Building the neighbor list
#> ----------------------------------------
#> ----------------------------------------
#> Building the neighbors of neighbors list
#> ----------------------------------------
#> ----------------------------------------
#> 	Model description
#> ----------------------------------------
#> NNGP Integrated Occupancy Model with Polya-Gamma latent
#> variable fit with 54 sites.
#> 
#> Integrating 4 occupancy data sets.
#> 
#> Number of MCMC samples: 1000 (40 batches of length 25)
#> Burn-in: 500 
#> Thinning Rate: 1 
#> Total Posterior Samples: 500 
#> 
#> Using the spherical spatial correlation model.
#> 
#> Using 5 nearest neighbors.
#> 
#> 
#> Source compiled with OpenMP support and model fit using 1 thread(s).
#> 
#> Adaptive Metropolis with target acceptance rate: 43.0
#> Sampling ... 
#> Batch: 10 of 40, 25.00%
#> 	Acceptance	Tuning
#> 	76.0		1.11628
#> -------------------------------------------------
#> Batch: 20 of 40, 50.00%
#> 	Acceptance	Tuning
#> 	72.0		1.23368
#> -------------------------------------------------
#> Batch: 30 of 40, 75.00%
#> 	Acceptance	Tuning
#> 	68.0		1.36343
#> -------------------------------------------------
#> Batch: 40 of 40, 100.00%
summary(out)
#> 
#> Call:
#> spIntPGOcc(occ.formula = ~occ.cov, det.formula = list(f.1 = ~det.cov.1.1, 
#>     f.2 = ~det.cov.2.1 + det.cov.2.2, f.3 = ~det.cov.3.1, f.4 = ~det.cov.4.1 + 
#>         det.cov.4.2 + det.cov.4.3), data = data.list, inits = inits.list, 
#>     priors = prior.list, tuning = tuning.list, cov.model = "spherical", 
#>     NNGP = TRUE, n.neighbors = 5, search.type = "cb", n.batch = n.batch, 
#>     batch.length = batch.length, accept.rate = 0.43, n.omp.threads = 1, 
#>     verbose = TRUE, n.report = 10, n.burn = 500, n.thin = 1)
#> 
#> Chain Information:
#> Total samples: 1000
#> Burn-in: 500
#> Thin: 1
#> Total Posterior Samples: 500
#> 
#> Occurrence: 
#>                2.5%    25%    50%    75%  97.5%
#> (Intercept) -0.2636  0.247 0.5129 0.8141 1.4424
#> occ.cov     -0.5966 -0.171 0.0874 0.3255 0.8338
#> 
#> Data source 1 Detection: 
#>                2.5%    25%    50%    75%  97.5%
#> (Intercept) -0.2282 0.4195 0.7745 1.2014 2.0673
#> det.cov.1.1  0.1654 0.7089 1.1333 1.5707 2.4026
#> 
#> Data source 2 Detection: 
#>                2.5%    25%    50%    75%  97.5%
#> (Intercept) -0.1516 0.2750 0.5198 0.7470 1.2763
#> det.cov.2.1  0.2214 0.7133 1.0141 1.3155 1.9901
#> det.cov.2.2  0.1417 0.5689 0.8220 1.1126 1.7367
#> 
#> Data source 3 Detection: 
#>                2.5%     25%     50%     75%  97.5%
#> (Intercept) -0.9968 -0.5536 -0.2994 -0.0250 0.4088
#> det.cov.3.1 -0.2799  0.1317  0.3648  0.5616 1.0190
#> 
#> Data source 4 Detection: 
#>                2.5%     25%     50%     75%   97.5%
#> (Intercept) -0.4467  0.1677  0.4109  0.7230  1.1684
#> det.cov.4.1 -0.5122 -0.0609  0.1500  0.4253  0.8894
#> det.cov.4.2 -2.0301 -1.4035 -1.0668 -0.7993 -0.3161
#> det.cov.4.3 -1.4359 -0.8659 -0.5623 -0.2788  0.3395
#> 
#> Covariance: 
#>            2.5%     25%     50%     75%   97.5%
#> sigma.sq 0.3812  0.7856  1.2757  2.4120  5.9378
#> phi      4.0679 10.1978 16.8131 22.6088 28.3439

# Predict at new locations ------------------------------------------------
out.pred <- predict(out, X.0, coords.0, verbose = FALSE)