Function for prediction at new locations for multispecies spatial occupancy models

The function predict collects posterior predictive samples for a set of new locations given an object of class `spMsPGOcc`.

# S3 method for spMsPGOcc
predict(object, X.0, coords.0, n.omp.threads = 1, verbose = TRUE, 
                            n.report = 100, ...)

Arguments

object
an object of class spMsPGOcc

X.0
the design matrix for prediction locations. This should include a column of 1s for the intercept. Covariates should have the same column names as those used when fitting the model with intPGOcc.

coords.0
the spatial coordinates corresponding to X.0.

n.omp.threads
a positive integer indicating the number of threads to use for SMP parallel processing. The package must be compiled for OpenMP support. For most Intel-based machines, we recommend setting n.omp.threads up to the number of hyperthreaded cores. Note, n.omp.threads > 1 might not work on some systems.

verbose
if TRUE, model specification and progress of the sampler is printed to the screen. Otherwise, nothing is printed to the screen.

n.report
the interval to report sampling progress.

...
currently no additional arguments

Author

Jeffrey W. Doser doserjef@msu.edu,
Andrew O. Finley finleya@msu.edu

Value

An object of class predict.spMsPGOcc that is a list comprised of:

psi.0.samples

a three-dimensional array of posterior predictive samples for the latent occurrence probability values.

z.0.samples

a three-dimensional array of posterior predictive samples for the latent occurrence values.

w.0.samples

a three-dimensional array of posterior predictive samples for the latent spatial random effects.

run.time

execution time reported using proc.time().

The return object will include additional objects used for standard extractor functions.
Examples

set.seed(400)

# Simulate Data -----------------------------------------------------------
J.x <- 7
J.y <- 7
J <- J.x * J.y
n.rep <- sample(2:4, size = J, replace = TRUE)
N <- 5
# Community-level covariate effects
# Occurrence
beta.mean <- c(0.2, -0.15)
p.occ <- length(beta.mean)
tau.sq.beta <- c(0.6, 0.3)
# Detection
alpha.mean <- c(0.5, 0.2, -.2)
tau.sq.alpha <- c(0.2, 0.3, 0.8)
p.det <- length(alpha.mean)
# Draw species-level effects from community means.
beta <- matrix(NA, nrow = N, ncol = p.occ)
alpha <- matrix(NA, nrow = N, ncol = p.det)
for (i in 1:p.occ) {
  beta[, i] <- rnorm(N, beta.mean[i], sqrt(tau.sq.beta[i]))
}
for (i in 1:p.det) {
  alpha[, i] <- rnorm(N, alpha.mean[i], sqrt(tau.sq.alpha[i]))
}
phi <- runif(N, 3/1, 3/.4)
sigma.sq <- runif(N, 0.3, 3)
sp <- TRUE

dat <- simMsOcc(J.x = J.x, J.y = J.y, n.rep = n.rep, N = N, beta = beta, alpha = alpha,
    phi = phi, sigma.sq = sigma.sq, sp = TRUE, cov.model = 'exponential')

# Number of batches
n.batch <- 40
# Batch length
batch.length <- 25
n.samples <- n.batch * batch.length

# Split into fitting and prediction data set
pred.indx <- sample(1:J, round(J * .25), replace = FALSE)
y <- dat$y[, -pred.indx, ]
# Occupancy covariates
X <- dat$X[-pred.indx, ]
# Coordinates
coords <- as.matrix(dat$coords[-pred.indx, ])
# Detection covariates
X.p <- dat$X.p[-pred.indx, , ]
# Prediction values
X.0 <- dat$X[pred.indx, ]
coords.0 <- as.matrix(dat$coords[pred.indx, ])
psi.0 <- dat$psi[, pred.indx]

# Package all data into a list
occ.covs <- X[, 2, drop = FALSE]
colnames(occ.covs) <- c('occ.cov')
det.covs <- list(det.cov.1 = X.p[, , 2], 
     det.cov.2 = X.p[, , 3]
     )
data.list <- list(y = y, 
      occ.covs = occ.covs,
      det.covs = det.covs, 
      coords = coords)

# Priors 
prior.list <- list(beta.comm.normal = list(mean = 0, var = 2.72), 
       alpha.comm.normal = list(mean = 0, var = 2.72), 
       tau.sq.beta.ig = list(a = 0.1, b = 0.1), 
       tau.sq.alpha.ig = list(a = 0.1, b = 0.1),
       phi.unif = list(a = 3/1, b = 3/.1), 
       sigma.sq.ig = list(a = 2, b = 2)) 
# Starting values
inits.list <- list(alpha.comm = 0, 
          beta.comm = 0, 
          beta = 0, 
          alpha = 0,
          tau.sq.beta = 1, 
          tau.sq.alpha = 1, 
          phi = 3 / .5, 
          sigma.sq = 2,
          w = matrix(0, nrow = N, ncol = nrow(X)),
          z = apply(y, c(1, 2), max, na.rm = TRUE))
# Tuning
tuning.list <- list(phi = 1) 

out <- spMsPGOcc(occ.formula = ~ occ.cov, 
     det.formula = ~ det.cov.1 + det.cov.2, 
     data = data.list,
     inits = inits.list, 
     n.batch = n.batch, 
     batch.length = batch.length, 
     accept.rate = 0.43, 
           priors = prior.list, 
     cov.model = "exponential", 
     tuning = tuning.list, 
           n.omp.threads = 1, 
           verbose = TRUE, 
     NNGP = TRUE, 
     n.neighbors = 5, 
     search.type = 'cb', 
           n.report = 10, 
     n.burn = 500, 
     n.thin = 1)
#> ----------------------------------------
#> 	Preparing the data
#> ----------------------------------------
#> ----------------------------------------
#> 	Building the neighbor list
#> ----------------------------------------
#> ----------------------------------------
#> Building the neighbors of neighbors list
#> ----------------------------------------
#> ----------------------------------------
#> 	Model description
#> ----------------------------------------
#> NNGP Multispecies Occupancy Model with Polya-Gamma latent
#> variable fit with 37 sites and 5 species.
#> 
#> Number of MCMC samples: 1000 (40 batches of length 25)
#> Burn-in: 500 
#> Thinning Rate: 1 
#> Total Posterior Samples: 500 
#> 
#> Using the exponential spatial correlation model.
#> 
#> Using 5 nearest neighbors.
#> 
#> Source compiled with OpenMP support and model fit using 1 thread(s).
#> 
#> Adaptive Metropolis with target acceptance rate: 43.0
#> Sampling ... 
#> Batch: 10 of 40, 25.00%
#> 	Species		Acceptance	Tuning
#> 	1		76.0		1.11628
#> 	2		80.0		1.11628
#> 	3		84.0		1.11628
#> 	4		68.0		1.11628
#> 	5		84.0		1.11628
#> -------------------------------------------------
#> Batch: 20 of 40, 50.00%
#> 	Species		Acceptance	Tuning
#> 	1		56.0		1.23368
#> 	2		84.0		1.23368
#> 	3		72.0		1.23368
#> 	4		76.0		1.23368
#> 	5		68.0		1.23368
#> -------------------------------------------------
#> Batch: 30 of 40, 75.00%
#> 	Species		Acceptance	Tuning
#> 	1		56.0		1.36343
#> 	2		56.0		1.36343
#> 	3		76.0		1.36343
#> 	4		64.0		1.36343
#> 	5		52.0		1.36343
#> -------------------------------------------------
#> Batch: 40 of 40, 100.00%

summary(out, level = 'both')
#> 
#> Call:
#> spMsPGOcc(occ.formula = ~occ.cov, det.formula = ~det.cov.1 + 
#>     det.cov.2, data = data.list, inits = inits.list, priors = prior.list, 
#>     tuning = tuning.list, cov.model = "exponential", NNGP = TRUE, 
#>     n.neighbors = 5, search.type = "cb", n.batch = n.batch, batch.length = batch.length, 
#>     accept.rate = 0.43, n.omp.threads = 1, verbose = TRUE, n.report = 10, 
#>     n.burn = 500, n.thin = 1)
#> 
#> Chain Information:
#> Total samples: 1000
#> Burn-in: 500
#> Thin: 1
#> Total Posterior Samples: 500
#> 
#> ----------------------------------------
#> 	Community Level
#> ----------------------------------------
#> Occurrence Means: 
#>                2.5%     25%     50%     75%  97.5%
#> (Intercept) -1.5837 -0.8883 -0.6228 -0.3586 0.2853
#> occ.cov     -0.7511 -0.1298  0.1439  0.3627 0.9433
#> 
#> Occurrence Variances: 
#>               2.5%    25%    50%    75%  97.5%
#> (Intercept) 0.0769 0.2988 0.5327 1.0680 5.8304
#> occ.cov     0.0501 0.2175 0.4730 0.8782 3.7879
#> 
#> Detection Means: 
#>                2.5%     25%     50%     75%  97.5%
#> (Intercept) -0.0701  0.5586  0.7462  0.9718 1.4710
#> det.cov.1   -0.6312 -0.2144  0.0116  0.1922 0.6940
#> det.cov.2   -1.1160 -0.6064 -0.4191 -0.2468 0.3456
#> 
#> Detection Variances: 
#>               2.5%    25%    50%    75%  97.5%
#> (Intercept) 0.0508 0.1656 0.3119 0.6047 2.3762
#> det.cov.1   0.0564 0.1671 0.2935 0.5507 2.3717
#> det.cov.2   0.0485 0.1380 0.2781 0.5623 2.9141
#> 
#> ----------------------------------------
#> 	Species Level
#> ----------------------------------------
#> Occurrence: 
#>                    2.5%     25%     50%     75%   97.5%
#> (Intercept)-sp1 -1.7280 -0.8482 -0.5385 -0.2364  0.4612
#> (Intercept)-sp2 -0.7710 -0.2376  0.0436  0.3443  0.9839
#> (Intercept)-sp3 -1.9519 -1.3136 -0.9580 -0.6627 -0.0909
#> (Intercept)-sp4 -1.3618 -0.7679 -0.4757 -0.1813  0.4200
#> (Intercept)-sp5 -2.6354 -1.7827 -1.3679 -0.9835 -0.4048
#> occ.cov-sp1     -1.6728 -0.8574 -0.4976 -0.2193  0.3758
#> occ.cov-sp2     -0.6771 -0.1633  0.0804  0.3234  0.8651
#> occ.cov-sp3     -0.6191 -0.0348  0.2984  0.6127  1.2210
#> occ.cov-sp4     -0.0141  0.4610  0.7677  1.0844  1.6832
#> occ.cov-sp5     -0.6061 -0.0101  0.2745  0.5100  1.0944
#> 
#> Detection: 
#>                    2.5%     25%     50%     75%   97.5%
#> (Intercept)-sp1  0.2789  0.6969  0.9273  1.1531  1.5821
#> (Intercept)-sp2  0.5148  0.8938  1.0905  1.3051  1.7628
#> (Intercept)-sp3 -0.8467 -0.2313  0.1363  0.4629  0.9732
#> (Intercept)-sp4  0.2605  0.6246  0.8281  1.0510  1.4072
#> (Intercept)-sp5 -0.0968  0.5181  0.7971  1.0595  1.6353
#> det.cov.1-sp1   -0.4421 -0.1342  0.0722  0.2834  0.6670
#> det.cov.1-sp2   -1.0458 -0.6514 -0.4241 -0.2522  0.0820
#> det.cov.1-sp3   -0.3559  0.0865  0.3143  0.5604  1.0799
#> det.cov.1-sp4   -0.1073  0.2143  0.3978  0.6404  1.0444
#> det.cov.1-sp5   -1.2277 -0.5863 -0.3116 -0.0505  0.4081
#> det.cov.2-sp1   -0.5935 -0.2252 -0.0253  0.1704  0.6205
#> det.cov.2-sp2   -0.9428 -0.6158 -0.4455 -0.2771  0.0457
#> det.cov.2-sp3   -2.0933 -1.2471 -0.9510 -0.6839 -0.2032
#> det.cov.2-sp4   -1.2956 -0.8699 -0.6530 -0.4567 -0.1464
#> det.cov.2-sp5   -0.7147 -0.3246 -0.0895  0.1449  0.6471
#> 
#> Covariance: 
#>                2.5%     25%     50%     75%   97.5%
#> sigma.sq-sp1 0.4059  0.7017  1.3781  3.3771 16.0847
#> sigma.sq-sp2 0.3611  0.7672  1.2161  2.1300  7.1000
#> sigma.sq-sp3 0.4684  0.9140  1.8019  3.3305  8.5747
#> sigma.sq-sp4 0.3627  0.6442  0.9404  1.5747  5.6841
#> sigma.sq-sp5 0.3470  0.7780  1.2907  2.0021  4.9853
#> phi-sp1      3.2295 10.9378 16.6678 23.9878 29.4061
#> phi-sp2      3.5035  9.2047 16.9292 22.2700 29.1258
#> phi-sp3      5.7048 14.0022 21.7802 26.1145 29.2951
#> phi-sp4      4.8799 10.8393 18.1445 24.9218 29.7763
#> phi-sp5      3.2460  7.2747 14.2834 22.6810 29.2382

# Predict at new locations ------------------------------------------------
out.pred <- predict(out, X.0, coords.0, verbose = FALSE)