Function for prediction at new locations for multi-species spatial occupancy models

The function predict collects posterior predictive samples for a set of new locations given an object of class `spMsPGOcc`.

# S3 method for spMsPGOcc
predict(object, X.0, coords.0, n.omp.threads = 1, verbose = TRUE, 
                            n.report = 100, ...)

Arguments

object

an object of class spMsPGOcc

X.0

the design matrix for prediction locations. This should include a column of 1s for the intercept. Covariates should have the same column names as those used when fitting the model with intPGOcc.

coords.0

the spatial coordinates corresponding to X.0.

n.omp.threads

a positive integer indicating the number of threads to use for SMP parallel processing. The package must be compiled for OpenMP support. For most Intel-based machines, we recommend setting n.omp.threads up to the number of hyperthreaded cores. Note, n.omp.threads > 1 might not work on some systems.

verbose

if TRUE, model specification and progress of the sampler is printed to the screen. Otherwise, nothing is printed to the screen.

n.report

the interval to report sampling progress.

...

currently no additional arguments

Author

Jeffrey W. Doser doserjef@msu.edu,
Andrew O. Finley finleya@msu.edu

Value

An object of class predict.spMsPGOcc that is a list comprised of:

psi.0.samples

a three-dimensional array of posterior predictive samples for the latent occurrence probability values.

z.0.samples

a three-dimensional array of posterior predictive samples for the latent occurrence values.

w.0.samples

a three-dimensional array of posterior predictive samples for the latent spatial random effects.

run.time

execution time reported using proc.time().

The return object will include additional objects used for standard extractor functions.

Examples

set.seed(400)

# Simulate Data -----------------------------------------------------------
J.x <- 7
J.y <- 7
J <- J.x * J.y
n.rep <- sample(2:4, size = J, replace = TRUE)
N <- 5
# Community-level covariate effects
# Occurrence
beta.mean <- c(0.2, -0.15)
p.occ <- length(beta.mean)
tau.sq.beta <- c(0.6, 0.3)
# Detection
alpha.mean <- c(0.5, 0.2, -.2)
tau.sq.alpha <- c(0.2, 0.3, 0.8)
p.det <- length(alpha.mean)
# Draw species-level effects from community means.
beta <- matrix(NA, nrow = N, ncol = p.occ)
alpha <- matrix(NA, nrow = N, ncol = p.det)
for (i in 1:p.occ) {
  beta[, i] <- rnorm(N, beta.mean[i], sqrt(tau.sq.beta[i]))
}
for (i in 1:p.det) {
  alpha[, i] <- rnorm(N, alpha.mean[i], sqrt(tau.sq.alpha[i]))
}
phi <- runif(N, 3/1, 3/.4)
sigma.sq <- runif(N, 0.3, 3)
sp <- TRUE

dat <- simMsOcc(J.x = J.x, J.y = J.y, n.rep = n.rep, N = N, beta = beta, alpha = alpha,
    phi = phi, sigma.sq = sigma.sq, sp = TRUE, cov.model = 'exponential')

# Number of batches
n.batch <- 30
# Batch length
batch.length <- 25
n.samples <- n.batch * batch.length

# Split into fitting and prediction data set
pred.indx <- sample(1:J, round(J * .25), replace = FALSE)
y <- dat$y[, -pred.indx, ]
# Occupancy covariates
X <- dat$X[-pred.indx, ]
# Coordinates
coords <- as.matrix(dat$coords[-pred.indx, ])
# Detection covariates
X.p <- dat$X.p[-pred.indx, , ]
# Prediction values
X.0 <- dat$X[pred.indx, ]
coords.0 <- as.matrix(dat$coords[pred.indx, ])
psi.0 <- dat$psi[, pred.indx]

# Package all data into a list
occ.covs <- X[, 2, drop = FALSE]
colnames(occ.covs) <- c('occ.cov')
det.covs <- list(det.cov.1 = X.p[, , 2], 
     det.cov.2 = X.p[, , 3]
     )
data.list <- list(y = y, 
      occ.covs = occ.covs,
      det.covs = det.covs, 
      coords = coords)

# Priors 
prior.list <- list(beta.comm.normal = list(mean = 0, var = 2.72), 
       alpha.comm.normal = list(mean = 0, var = 2.72), 
       tau.sq.beta.ig = list(a = 0.1, b = 0.1), 
       tau.sq.alpha.ig = list(a = 0.1, b = 0.1),
       phi.unif = list(a = 3/1, b = 3/.1), 
       sigma.sq.ig = list(a = 2, b = 2)) 
# Starting values
inits.list <- list(alpha.comm = 0, 
          beta.comm = 0, 
          beta = 0, 
          alpha = 0,
          tau.sq.beta = 1, 
          tau.sq.alpha = 1, 
          phi = 3 / .5, 
          sigma.sq = 2,
          w = matrix(0, nrow = N, ncol = nrow(X)),
          z = apply(y, c(1, 2), max, na.rm = TRUE))
# Tuning
tuning.list <- list(phi = 1) 

out <- spMsPGOcc(occ.formula = ~ occ.cov, 
     det.formula = ~ det.cov.1 + det.cov.2, 
     data = data.list,
     inits = inits.list, 
     n.batch = n.batch, 
     batch.length = batch.length, 
     accept.rate = 0.43, 
           priors = prior.list, 
     cov.model = "exponential", 
     tuning = tuning.list, 
           n.omp.threads = 1, 
           verbose = TRUE, 
     NNGP = TRUE, 
     n.neighbors = 5, 
     search.type = 'cb', 
           n.report = 10, 
     n.burn = 500, 
     n.thin = 1)
#> ----------------------------------------
#> 	Preparing the data
#> ----------------------------------------
#> ----------------------------------------
#> 	Building the neighbor list
#> ----------------------------------------
#> ----------------------------------------
#> Building the neighbors of neighbors list
#> ----------------------------------------
#> ----------------------------------------
#> 	Model description
#> ----------------------------------------
#> NNGP Multispecies Occupancy Model with Polya-Gamma latent
#> variable fit with 37 sites and 5 species.
#> 
#> Samples per chain: 750 (30 batches of length 25)
#> Burn-in: 500 
#> Thinning Rate: 1 
#> Number of Chains: 1 
#> Total Posterior Samples: 250 
#> 
#> Using the exponential spatial correlation model.
#> 
#> Using 5 nearest neighbors.
#> 
#> Source compiled with OpenMP support and model fit using 1 thread(s).
#> 
#> Adaptive Metropolis with target acceptance rate: 43.0
#> ----------------------------------------
#> 	Chain 1
#> ----------------------------------------
#> Sampling ... 
#> Batch: 10 of 30, 33.33%
#> 	Species		Acceptance	Tuning
#> 	1		76.0		1.11628
#> 	2		80.0		1.11628
#> 	3		84.0		1.11628
#> 	4		68.0		1.11628
#> 	5		84.0		1.11628
#> -------------------------------------------------
#> Batch: 20 of 30, 66.67%
#> 	Species		Acceptance	Tuning
#> 	1		56.0		1.23368
#> 	2		84.0		1.23368
#> 	3		72.0		1.23368
#> 	4		76.0		1.23368
#> 	5		68.0		1.23368
#> -------------------------------------------------
#> Batch: 30 of 30, 100.00%

summary(out, level = 'both')
#> 
#> Call:
#> spMsPGOcc(occ.formula = ~occ.cov, det.formula = ~det.cov.1 + 
#>     det.cov.2, data = data.list, inits = inits.list, priors = prior.list, 
#>     tuning = tuning.list, cov.model = "exponential", NNGP = TRUE, 
#>     n.neighbors = 5, search.type = "cb", n.batch = n.batch, batch.length = batch.length, 
#>     accept.rate = 0.43, n.omp.threads = 1, verbose = TRUE, n.report = 10, 
#>     n.burn = 500, n.thin = 1)
#> 
#> Samples per Chain: 750
#> Burn-in: 500
#> Thinning Rate: 1
#> Number of Chains: 1
#> Total Posterior Samples: 250
#> Run Time (min): 0.0187
#> 
#> ----------------------------------------
#> 	Community Level
#> ----------------------------------------
#> Occurrence Means (logit scale): 
#>                Mean     SD    2.5%     50%  97.5% Rhat      ESS
#> (Intercept) -0.6759 0.4567 -1.6198 -0.6460 0.1321   NA  67.0536
#> occ.cov      0.0872 0.4315 -0.7382  0.0889 0.8809   NA 101.6429
#> 
#> Occurrence Variances (logit scale): 
#>               Mean     SD   2.5%    50%  97.5% Rhat     ESS
#> (Intercept) 0.9050 1.5427 0.0614 0.4185 4.9761   NA 64.2983
#> occ.cov     0.7879 1.1641 0.0656 0.4501 3.6613   NA 84.8482
#> 
#> Detection Means (logit scale): 
#>                Mean     SD    2.5%     50%  97.5% Rhat      ESS
#> (Intercept)  0.7549 0.3499 -0.0308  0.7641 1.4291   NA 125.1103
#> det.cov.1   -0.0217 0.3509 -0.7357 -0.0087 0.6906   NA 261.5600
#> det.cov.2   -0.4513 0.3655 -1.1177 -0.4578 0.2424   NA 108.6137
#> 
#> Detection Variances (logit scale): 
#>               Mean     SD   2.5%    50%  97.5% Rhat      ESS
#> (Intercept) 0.5096 1.0190 0.0470 0.2694 2.1767   NA 125.3159
#> det.cov.1   0.5542 0.8146 0.0641 0.3202 2.6724   NA  91.5808
#> det.cov.2   0.5369 1.2002 0.0632 0.2864 2.2280   NA 134.2719
#> 
#> ----------------------------------------
#> 	Species Level
#> ----------------------------------------
#> Occurrence (logit scale): 
#>                    Mean     SD    2.5%     50%   97.5% Rhat      ESS
#> (Intercept)-sp1 -0.6040 0.4925 -1.7339 -0.5470  0.2207   NA  31.4859
#> (Intercept)-sp2 -0.0437 0.4164 -0.6982 -0.0659  0.8529   NA  84.4176
#> (Intercept)-sp3 -1.0377 0.5085 -2.1132 -0.9940 -0.1497   NA  60.9916
#> (Intercept)-sp4 -0.5636 0.4759 -1.6050 -0.5658  0.3288   NA  74.9439
#> (Intercept)-sp5 -1.3000 0.5762 -2.5957 -1.2561 -0.3540   NA  29.5920
#> occ.cov-sp1     -0.5667 0.5059 -1.7576 -0.5106  0.3353   NA  53.2553
#> occ.cov-sp2      0.0474 0.3953 -0.6933  0.0580  0.8536   NA 132.7523
#> occ.cov-sp3      0.2694 0.4489 -0.5674  0.2652  1.1739   NA  77.9782
#> occ.cov-sp4      0.8004 0.4795 -0.0796  0.8184  1.7700   NA  67.5391
#> occ.cov-sp5      0.2136 0.4151 -0.6179  0.2257  1.0708   NA 120.4539
#> 
#> Detection (logit scale): 
#>                    Mean     SD    2.5%     50%   97.5% Rhat      ESS
#> (Intercept)-sp1  0.9556 0.3304  0.3902  0.9200  1.5778   NA 133.4310
#> (Intercept)-sp2  1.1192 0.3089  0.5059  1.0949  1.7552   NA 121.1634
#> (Intercept)-sp3  0.2020 0.4352 -0.6586  0.2305  1.0098   NA  39.1465
#> (Intercept)-sp4  0.8554 0.2980  0.3345  0.8440  1.3892   NA 115.7020
#> (Intercept)-sp5  0.7765 0.4286 -0.0120  0.7977  1.5970   NA 175.7905
#> det.cov.1-sp1    0.0977 0.3097 -0.4241  0.0945  0.7278   NA 174.1432
#> det.cov.1-sp2   -0.4643 0.3005 -1.1227 -0.4341  0.0782   NA 118.7330
#> det.cov.1-sp3    0.3084 0.3634 -0.3717  0.2863  1.0280   NA  84.5381
#> det.cov.1-sp4    0.4041 0.3269 -0.1320  0.3696  1.0420   NA  86.8029
#> det.cov.1-sp5   -0.3866 0.4031 -1.2316 -0.3540  0.2440   NA 123.4658
#> det.cov.2-sp1   -0.0289 0.3060 -0.6395 -0.0277  0.6134   NA 115.9287
#> det.cov.2-sp2   -0.4488 0.2403 -0.9266 -0.4443  0.0188   NA 177.0864
#> det.cov.2-sp3   -1.0503 0.4393 -1.9894 -1.0133 -0.3869   NA  74.1371
#> det.cov.2-sp4   -0.7273 0.3076 -1.3230 -0.7026 -0.1677   NA 124.0647
#> det.cov.2-sp5   -0.0867 0.3450 -0.6582 -0.0917  0.6818   NA 123.9276
#> 
#> Spatial Covariance: 
#>                 Mean     SD   2.5%     50%   97.5% Rhat     ESS
#> sigma.sq-sp1  2.6745 4.2962 0.3721  1.0895 16.2718   NA  8.7219
#> sigma.sq-sp2  1.4410 1.1379 0.3371  1.0837  5.4434   NA 15.9288
#> sigma.sq-sp3  2.9950 2.2685 0.5632  2.3258  8.9187   NA 18.7938
#> sigma.sq-sp4  1.7848 1.7660 0.3565  1.0854  7.1363   NA 12.1886
#> sigma.sq-sp5  1.7337 1.4266 0.3476  1.2722  6.0217   NA 18.0503
#> phi-sp1      18.7757 8.3759 4.7999 19.1111 29.6378   NA 15.1445
#> phi-sp2      16.1790 7.6507 3.2580 16.7429 29.4304   NA 13.5254
#> phi-sp3      19.8704 6.7022 6.9681 21.7451 28.8376   NA 35.3363
#> phi-sp4      19.7948 7.2499 4.2368 20.9777 29.9211   NA 22.6146
#> phi-sp5      15.2747 7.7022 3.5589 14.6840 28.5104   NA 20.5954

# Predict at new locations ------------------------------------------------
out.pred <- predict(out, X.0, coords.0, verbose = FALSE)