Function for prediction at new locations for multi-species spatial occupancy models

The function predict collects posterior predictive samples for a set of new locations given an object of class `spMsPGOcc`. Prediction is possible for both the latent occupancy state as well as detection.

Usage

# S3 method for spMsPGOcc
predict(object, X.0, coords.0, n.omp.threads = 1, verbose = TRUE, 
                            n.report = 100, ignore.RE = FALSE, type = 'occupancy', ...)

Arguments

object

an object of class spMsPGOcc

X.0

the design matrix of covariates at the prediction locations. This should include a column of 1s for the intercept if an intercept is included in the model. If random effects are included in the occupancy (or detection if type = 'detection') portion of the model, the levels of the random effects at the new locations should be included as a column in the design matrix. The ordering of the levels should match the ordering used to fit the data in spMsPGOcc. Columns should correspond to the order of how covariates were specified in the corresponding formula argument of spMsPGOcc. Column names of the random effects must match the name of the random effects, if specified in the corresponding formula argument of spMsPGOcc.

coords.0

the spatial coordinates corresponding to X.0. Note that spOccupancy assumes coordinates are specified in a projected coordinate system.

n.omp.threads

a positive integer indicating the number of threads to use for SMP parallel processing. The package must be compiled for OpenMP support. For most Intel-based machines, we recommend setting n.omp.threads up to the number of hyperthreaded cores. Note, n.omp.threads > 1 might not work on some systems.

verbose

if TRUE, model specification and progress of the sampler is printed to the screen. Otherwise, nothing is printed to the screen.

n.report

the interval to report sampling progress.

ignore.RE

a logical value indicating whether to include unstructured random effects for prediction. If TRUE, unstructured random effects will be ignored and prediction will only use the fixed effects and the spatial random effects. If FALSE, random effects will be included in the prediction for both observed and unobserved levels of the unstructured random effects.

type

a quoted keyword indicating what type of prediction to produce. Valid keywords are 'occupancy' to predict latent occupancy probability and latent occupancy values (this is the default), or 'detection' to predict detection probability given new values of detection covariates.

...

currently no additional arguments

Note

When ignore.RE = FALSE, both sampled levels and non-sampled levels of random effects are supported for prediction. For sampled levels, the posterior distribution for the random intercept corresponding to that level of the random effect will be used in the prediction. For non-sampled levels, random values are drawn from a normal distribution using the posterior samples of the random effect variance, which results in fully propagated uncertainty in predictions with models that incorporate random effects.

Author

Jeffrey W. Doser doserjef@msu.edu,
Andrew O. Finley finleya@msu.edu

Value

An list object of class predict.spMsPGOcc. When type = 'occupancy', the list consists of:

psi.0.samples

a three-dimensional array of posterior predictive samples for the latent occurrence probability values.

z.0.samples

a three-dimensional array of posterior predictive samples for the latent occurrence values.

w.0.samples

a three-dimensional array of posterior predictive samples for the latent spatial random effects.

run.time

execution time reported using proc.time().

When type = 'detection', the list consists of:

p.0.samples

a three-dimensional array of posterior predictive samples for the detection probability values.

run.time

execution time reported using proc.time().

The return object will include additional objects used for standard extractor functions.

Examples

set.seed(400)

# Simulate Data -----------------------------------------------------------
J.x <- 7
J.y <- 7
J <- J.x * J.y
n.rep <- sample(2:4, size = J, replace = TRUE)
N <- 5
# Community-level covariate effects
# Occurrence
beta.mean <- c(0.2, -0.15)
p.occ <- length(beta.mean)
tau.sq.beta <- c(0.6, 0.3)
# Detection
alpha.mean <- c(0.5, 0.2, -.2)
tau.sq.alpha <- c(0.2, 0.3, 0.8)
p.det <- length(alpha.mean)
# Draw species-level effects from community means.
beta <- matrix(NA, nrow = N, ncol = p.occ)
alpha <- matrix(NA, nrow = N, ncol = p.det)
for (i in 1:p.occ) {
  beta[, i] <- rnorm(N, beta.mean[i], sqrt(tau.sq.beta[i]))
}
for (i in 1:p.det) {
  alpha[, i] <- rnorm(N, alpha.mean[i], sqrt(tau.sq.alpha[i]))
}
phi <- runif(N, 3/1, 3/.4)
sigma.sq <- runif(N, 0.3, 3)
sp <- TRUE

dat <- simMsOcc(J.x = J.x, J.y = J.y, n.rep = n.rep, N = N, beta = beta, alpha = alpha,
    phi = phi, sigma.sq = sigma.sq, sp = TRUE, cov.model = 'exponential')

# Number of batches
n.batch <- 30
# Batch length
batch.length <- 25
n.samples <- n.batch * batch.length

# Split into fitting and prediction data set
pred.indx <- sample(1:J, round(J * .25), replace = FALSE)
y <- dat$y[, -pred.indx, ]
# Occupancy covariates
X <- dat$X[-pred.indx, ]
# Coordinates
coords <- as.matrix(dat$coords[-pred.indx, ])
# Detection covariates
X.p <- dat$X.p[-pred.indx, , ]
# Prediction values
X.0 <- dat$X[pred.indx, ]
coords.0 <- as.matrix(dat$coords[pred.indx, ])
psi.0 <- dat$psi[, pred.indx]

# Package all data into a list
occ.covs <- X[, 2, drop = FALSE]
colnames(occ.covs) <- c('occ.cov')
det.covs <- list(det.cov.1 = X.p[, , 2], 
     det.cov.2 = X.p[, , 3]
     )
data.list <- list(y = y, 
      occ.covs = occ.covs,
      det.covs = det.covs, 
      coords = coords)

# Priors 
prior.list <- list(beta.comm.normal = list(mean = 0, var = 2.72), 
       alpha.comm.normal = list(mean = 0, var = 2.72), 
       tau.sq.beta.ig = list(a = 0.1, b = 0.1), 
       tau.sq.alpha.ig = list(a = 0.1, b = 0.1),
       phi.unif = list(a = 3/1, b = 3/.1), 
       sigma.sq.ig = list(a = 2, b = 2)) 
# Starting values
inits.list <- list(alpha.comm = 0, 
          beta.comm = 0, 
          beta = 0, 
          alpha = 0,
          tau.sq.beta = 1, 
          tau.sq.alpha = 1, 
          phi = 3 / .5, 
          sigma.sq = 2,
          w = matrix(0, nrow = N, ncol = nrow(X)),
          z = apply(y, c(1, 2), max, na.rm = TRUE))
# Tuning
tuning.list <- list(phi = 1) 

out <- spMsPGOcc(occ.formula = ~ occ.cov, 
     det.formula = ~ det.cov.1 + det.cov.2, 
     data = data.list,
     inits = inits.list, 
     n.batch = n.batch, 
     batch.length = batch.length, 
     accept.rate = 0.43, 
           priors = prior.list, 
     cov.model = "exponential", 
     tuning = tuning.list, 
           n.omp.threads = 1, 
           verbose = TRUE, 
     NNGP = TRUE, 
     n.neighbors = 5, 
     search.type = 'cb', 
           n.report = 10, 
     n.burn = 500, 
     n.thin = 1)
#> ----------------------------------------
#> 	Preparing to run the model
#> ----------------------------------------
#> ----------------------------------------
#> 	Building the neighbor list
#> ----------------------------------------
#> ----------------------------------------
#> Building the neighbors of neighbors list
#> ----------------------------------------
#> ----------------------------------------
#> 	Model description
#> ----------------------------------------
#> NNGP Multispecies Occupancy Model with Polya-Gamma latent
#> variable fit with 37 sites and 5 species.
#> 
#> Samples per chain: 750 (30 batches of length 25)
#> Burn-in: 500 
#> Thinning Rate: 1 
#> Number of Chains: 1 
#> Total Posterior Samples: 250 
#> 
#> Using the exponential spatial correlation model.
#> 
#> Using 5 nearest neighbors.
#> 
#> Source compiled with OpenMP support and model fit using 1 thread(s).
#> 
#> Adaptive Metropolis with target acceptance rate: 43.0
#> ----------------------------------------
#> 	Chain 1
#> ----------------------------------------
#> Sampling ... 
#> Batch: 10 of 30, 33.33%
#> 	Species		Parameter	Acceptance	Tuning
#> 	1		phi		60.0		1.11628
#> 	2		phi		76.0		1.09417
#> 	3		phi		76.0		1.11628
#> 	4		phi		76.0		1.11628
#> 	5		phi		48.0		1.11628
#> -------------------------------------------------
#> Batch: 20 of 30, 66.67%
#> 	Species		Parameter	Acceptance	Tuning
#> 	1		phi		76.0		1.23368
#> 	2		phi		80.0		1.20925
#> 	3		phi		80.0		1.20925
#> 	4		phi		72.0		1.23368
#> 	5		phi		64.0		1.23368
#> -------------------------------------------------
#> Batch: 30 of 30, 100.00%

summary(out, level = 'both')
#> 
#> Call:
#> spMsPGOcc(occ.formula = ~occ.cov, det.formula = ~det.cov.1 + 
#>     det.cov.2, data = data.list, inits = inits.list, priors = prior.list, 
#>     tuning = tuning.list, cov.model = "exponential", NNGP = TRUE, 
#>     n.neighbors = 5, search.type = "cb", n.batch = n.batch, batch.length = batch.length, 
#>     accept.rate = 0.43, n.omp.threads = 1, verbose = TRUE, n.report = 10, 
#>     n.burn = 500, n.thin = 1)
#> 
#> Samples per Chain: 750
#> Burn-in: 500
#> Thinning Rate: 1
#> Number of Chains: 1
#> Total Posterior Samples: 250
#> Run Time (min): 0.0202
#> 
#> ----------------------------------------
#> 	Community Level
#> ----------------------------------------
#> Occurrence Means (logit scale): 
#>                Mean     SD    2.5%     50%  97.5% Rhat ESS
#> (Intercept) -0.6273 0.4545 -1.4014 -0.6562 0.2345   NA  98
#> occ.cov      0.1919 0.3984 -0.5555  0.1651 0.9640   NA 141
#> 
#> Occurrence Variances (logit scale): 
#>              Mean     SD   2.5%    50%  97.5% Rhat ESS
#> (Intercept) 0.969 2.0574 0.0888 0.5308 3.5194   NA  85
#> occ.cov     0.848 1.8906 0.0532 0.3804 3.1291   NA 177
#> 
#> Detection Means (logit scale): 
#>                Mean     SD    2.5%     50%  97.5% Rhat ESS
#> (Intercept)  0.7244 0.3664 -0.0371  0.7351 1.3010   NA 114
#> det.cov.1    0.0291 0.4093 -0.8174  0.0173 0.8108   NA 159
#> det.cov.2   -0.4459 0.3144 -1.0905 -0.4327 0.1040   NA  98
#> 
#> Detection Variances (logit scale): 
#>               Mean     SD   2.5%    50%  97.5% Rhat ESS
#> (Intercept) 0.6667 2.3170 0.0477 0.2925 2.6475   NA 182
#> det.cov.1   0.6595 0.9138 0.0920 0.3770 2.8361   NA 134
#> det.cov.2   0.4938 0.6400 0.0489 0.3017 2.1889   NA 129
#> 
#> ----------------------------------------
#> 	Species Level
#> ----------------------------------------
#> Occurrence (logit scale): 
#>                    Mean     SD    2.5%     50%   97.5% Rhat ESS
#> (Intercept)-sp1 -0.5128 0.4568 -1.3238 -0.5101  0.3241   NA 108
#> (Intercept)-sp2 -0.0039 0.5472 -1.1124 -0.0240  1.0155   NA  40
#> (Intercept)-sp3 -0.8077 0.5926 -1.8009 -0.8288  0.6058   NA  49
#> (Intercept)-sp4 -0.5163 0.3985 -1.2227 -0.5575  0.2458   NA  47
#> (Intercept)-sp5 -1.4790 0.6285 -2.7975 -1.4236 -0.3900   NA  41
#> occ.cov-sp1     -0.4548 0.4454 -1.3704 -0.4334  0.3480   NA  81
#> occ.cov-sp2      0.1080 0.4087 -0.7317  0.1145  0.9183   NA 101
#> occ.cov-sp3      0.3354 0.4903 -0.4883  0.2933  1.3509   NA  71
#> occ.cov-sp4      0.8040 0.4672  0.0099  0.7858  1.7312   NA  74
#> occ.cov-sp5      0.2429 0.4509 -0.6938  0.2687  1.0806   NA  84
#> 
#> Detection (logit scale): 
#>                    Mean     SD    2.5%     50%   97.5% Rhat ESS
#> (Intercept)-sp1  0.9705 0.3130  0.3498  0.9793  1.5378   NA  97
#> (Intercept)-sp2  1.1200 0.2856  0.5849  1.1179  1.6843   NA 108
#> (Intercept)-sp3  0.1262 0.5219 -0.9213  0.1859  0.9642   NA  40
#> (Intercept)-sp4  0.8355 0.3206  0.2209  0.8554  1.4644   NA 108
#> (Intercept)-sp5  0.7792 0.4863 -0.2022  0.8405  1.7494   NA  63
#> det.cov.1-sp1    0.0428 0.3123 -0.6075  0.0627  0.6492   NA 181
#> det.cov.1-sp2   -0.4691 0.2707 -0.9692 -0.4559  0.0404   NA 108
#> det.cov.1-sp3    0.4341 0.3698 -0.2442  0.4063  1.1347   NA 128
#> det.cov.1-sp4    0.4530 0.2947 -0.0930  0.4459  1.0829   NA 145
#> det.cov.1-sp5   -0.3191 0.4261 -1.2088 -0.3139  0.4029   NA 121
#> det.cov.2-sp1   -0.0417 0.2951 -0.5540 -0.0633  0.6172   NA  91
#> det.cov.2-sp2   -0.4640 0.2510 -0.9161 -0.4473  0.0589   NA  97
#> det.cov.2-sp3   -0.9504 0.4117 -1.8531 -0.9175 -0.2875   NA  77
#> det.cov.2-sp4   -0.6669 0.2794 -1.2285 -0.6526 -0.2053   NA  93
#> det.cov.2-sp5   -0.0899 0.3173 -0.6740 -0.1021  0.5920   NA 192
#> 
#> Spatial Covariance: 
#>                 Mean      SD   2.5%     50%   97.5% Rhat ESS
#> sigma.sq-sp1  1.7912  1.8878 0.3248  1.1624  7.3931   NA  20
#> sigma.sq-sp2  2.4424  1.5781 0.5389  2.0426  6.6044   NA  18
#> sigma.sq-sp3  9.2552 14.4259 0.3857  2.4765 52.6109   NA   3
#> sigma.sq-sp4  1.3352  0.7837 0.4124  1.1462  3.4119   NA  25
#> sigma.sq-sp5  2.4297  2.9247 0.4503  1.4167  9.4182   NA  15
#> phi-sp1      16.8199  7.6410 3.8762 15.9487 29.7217   NA  14
#> phi-sp2      14.4504  8.2952 3.7476 12.9992 29.3505   NA  11
#> phi-sp3      18.9227  6.4968 6.8818 18.6367 28.9020   NA  56
#> phi-sp4      17.8393  7.9783 5.5983 17.2270 29.5593   NA  12
#> phi-sp5      14.1952  7.2692 3.4550 13.7622 27.0093   NA  20

# Predict at new locations ------------------------------------------------
out.pred <- predict(out, X.0, coords.0, verbose = FALSE)