Function for prediction at new locations for single species spatial occupancy models

The function predict collects posterior predictive samples for a set of new locations given an object of class `spPGOcc`.

# S3 method for spPGOcc
predict(object, X.0, coords.0, n.omp.threads = 1, verbose = TRUE, 
                          n.report = 100, ...)

Arguments

object
an object of class spPGOcc

X.0
the design matrix for prediction locations. This should include a column of 1s for the intercept if an intercept is included in the original model fit. Covariates should have the same column names as those used when fitting the model with spPGOcc.

coords.0
the spatial coordinates corresponding to X.0.

n.omp.threads
a positive integer indicating the number of threads to use for SMP parallel processing. The package must be compiled for OpenMP support. For most Intel-based machines, we recommend setting n.omp.threads up to the number of hyperthreaded cores. Note, n.omp.threads > 1 might not work on some systems.

verbose
if TRUE, model specification and progress of the sampler is printed to the screen. Otherwise, nothing is printed to the screen.

n.report
the interval to report sampling progress.

...
currently no additional arguments

Author

Jeffrey W. Doser doserjef@msu.edu,
Andrew O. Finley finleya@msu.edu

References

Hooten, M. B., and Hefley, T. J. (2019). Bringing Bayesian models to life. CRC Press.

Value

An object of class predict.spPGOcc that is a list comprised of:

psi.0.samples

a coda object of posterior predictive samples for the latent occurrence probability values.

z.0.samples

a coda object of posterior predictive samples for the latent occurrence values.

w.0.samples

a coda object of posterior predictive samples for the latent spatial random effects.

run.time

execution time reported using proc.time().

The return object will include additional objects used for standard extractor functions.
Examples

if (FALSE) {
set.seed(400)
# Simulate Data -----------------------------------------------------------
J.x <- 10
J.y <- 10
J <- J.x * J.y
n.rep <- sample(2:4, J, replace = TRUE)
beta <- c(0.5, 2)
p.occ <- length(beta)
alpha <- c(0, 1)
p.det <- length(alpha)
phi <- 3 / .6
sigma.sq <- 2
dat <- simOcc(J.x = J.x, J.y = J.y, n.rep = n.rep, beta = beta, alpha = alpha, 
              sigma.sq = sigma.sq, phi = phi, sp = TRUE)
# Split into fitting and prediction data set
pred.indx <- sample(1:J, round(J * .5), replace = FALSE)
y <- dat$y[-pred.indx, ]
# Occupancy covariates
X <- dat$X[-pred.indx, ]
# Prediction covariates
X.0 <- dat$X[pred.indx, ]
# Detection covariates
X.p <- dat$X.p[-pred.indx, , ]
coords <- as.matrix(dat$coords[-pred.indx, ])
coords.0 <- as.matrix(dat$coords[pred.indx, ])
psi.0 <- dat$psi[pred.indx]
w.0 <- dat$w[pred.indx]

# Package all data into a list
occ.covs <- X[, -1, drop = FALSE]
colnames(occ.covs) <- c('occ.cov')
det.covs <- list(det.cov.1 = X.p[, , 2])
data.list <- list(y = y, 
      occ.covs = occ.covs, 
      det.covs = det.covs, 
      coords = coords)

# Number of batches
n.batch <- 200
# Batch length
batch.length <- 25
n.iter <- n.batch * batch.length
# Priors 
prior.list <- list(beta.normal = list(mean = 0, var = 2.72), 
       alpha.normal = list(mean = 0, var = 2.72),
       sigma.sq.ig = c(2, 2), 
       phi.unif = c(3/1, 3/.1)) 
# Starting values
starting.list <- list(alpha = 0, beta = 0,
          phi = 3 / .5, 
          sigma.sq = 2,
          w = rep(0, nrow(X)),
          z = apply(y, 1, max, na.rm = TRUE))
# Tuning
tuning.list <- list(phi = 1) 

out <- spPGOcc(occ.formula = ~ occ.cov, 
         det.formula = ~ det.cov.1, 
         data = data.list, 
         starting = starting.list, 
         n.batch = n.batch, 
         batch.length = batch.length, 
         accept.rate = 0.43, 
         priors = prior.list,
         cov.model = 'exponential', 
         tuning = tuning.list, 
         n.omp.threads = 1, 
         verbose = TRUE, 
         NNGP = TRUE, 
         n.neighbors = 15, 
         search.type = 'cb', 
         n.report = 50)

summary(out) 

# Predict at new locations ------------------------------------------------
out.pred <- predict(out, X.0, coords.0, verbose = TRUE)
psi.0.quants <- apply(out.pred$psi.0.samples, 2, quantile, c(0.025, 0.5, 0.975))

# Plot of predicted psi vs true psi
# Code for plot is adapted from Hooten and Hefley (2019)
sq.cex=2.1
par(cex.lab=1.5,cex.main=1.5,xaxt="n",yaxt="n",bty="n",mar=c(5,5,4,2))
layout(matrix(1:2,1,2))
plot(coords.0,type="n",xlab="",ylab="",asp=TRUE,main="TRUE")
points(coords.0,pch=15,cex=sq.cex,col=rgb(0,0,0,(psi.0-min(psi.0))/
            diff(range(psi.0))))
plot(coords.0,type="n",xlab="",ylab="",asp=TRUE,main="PREDICTED")
points(coords.0,pch=15,cex=sq.cex,col=rgb(0,0,0,(psi.0.quants[2, ]-
             min(psi.0.quants[2, ]))/
                                          diff(range(psi.0.quants[2, ]))))
}